+++ pyproject.toml
@@ -12,7 +12,7 @@ maintainers = [
 ]
 license = {text = "Apache Software License"}
 dependencies = ["pandas<=2.2.1,>=0.25.0", "numpy<=1.26.4,>=1.22.0", "ordered-set<=4.1.0,>=4.0.2", "fugue<=0.8.7,>=0.8.7"]
-requires-python = ">=3.8.0"
+requires-python = ">=3.9.0"
 classifiers = [
     "Intended Audience :: Developers",
     "Natural Language :: English",
@@ -20,7 +20,6 @@ classifiers = [
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
