+++ tests/test_spark.py
@@ -13,2103 +13,1311 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import datetime
-import io
+"""
+Testing out the datacompy functionality
+"""
+
 import logging
 import re
+import sys
+from datetime import datetime
 from decimal import Decimal
+from io import StringIO
+from unittest import mock
 
+import numpy as np
+import pandas as pd
 import pytest
+from pytest import raises
 
 pytest.importorskip("pyspark")
 
-from pyspark.sql import Row, SparkSession
-from pyspark.sql.types import (
-    DateType,
-    DecimalType,
-    DoubleType,
-    LongType,
-    StringType,
-    StructField,
-    StructType,
+import pyspark.pandas as ps  # noqa: E402
+from pandas.testing import assert_series_equal  # noqa: E402
+
+from datacompy.spark import (  # noqa: E402
+    SparkCompare,
+    calculate_max_diff,
+    columns_equal,
+    generate_id_within_group,
+    temp_column_name,
 )
 
-import datacompy
-from datacompy import SparkCompare
-from datacompy.spark import _is_comparable
-
-# Turn off py4j debug messages for all tests in this module
-logging.getLogger("py4j").setLevel(logging.INFO)
-
-CACHE_INTERMEDIATES = True
+logging.basicConfig(stream=sys.stdout, level=logging.DEBUG)
 
 
-# Declare fixtures
-# (if we need to use these in other modules, move to conftest.py)
-@pytest.fixture(scope="module", name="spark")
-def spark_fixture():
-    spark = (
-        SparkSession.builder.master("local[2]")
-        .config("spark.driver.bindAddress", "127.0.0.1")
-        .appName("pytest")
-        .getOrCreate()
-    )
-    yield spark
-    spark.stop()
+pandas_version = pytest.mark.skipif(
+    pd.__version__ >= "2.0.0", reason="Pandas 2 is currently not supported"
+)
+
+pd.DataFrame.iteritems = pd.DataFrame.items  # Pandas 2+ compatability
+np.bool = np.bool_  # Numpy 1.24.3+ comptability
 
 
-@pytest.fixture(scope="module", name="base_df1")
-def base_df1_fixture(spark):
-    mock_data = [
-        Row(
-            acct=10000001234,
-            dollar_amt=123,
-            name="George Maharis",
-            float_fld=14530.1555,
-            date_fld=datetime.date(2017, 1, 1),
-        ),
-        Row(
-            acct=10000001235,
-            dollar_amt=0,
-            name="Michael Bluth",
-            float_fld=1.0,
-            date_fld=datetime.date(2017, 1, 1),
-        ),
-        Row(
-            acct=10000001236,
-            dollar_amt=1345,
-            name="George Bluth",
-            float_fld=None,
-            date_fld=datetime.date(2017, 1, 1),
-        ),
-        Row(
-            acct=10000001237,
-            dollar_amt=123456,
-            name="Bob Loblaw",
-            float_fld=345.12,
-            date_fld=datetime.date(2017, 1, 1),
-        ),
-        Row(
-            acct=10000001239,
-            dollar_amt=1,
-            name="Lucille Bluth",
-            float_fld=None,
-            date_fld=datetime.date(2017, 1, 1),
-        ),
-    ]
+@pandas_version
+def test_numeric_columns_equal_abs():
+    data = """a|b|expected
+1|1|True
+2|2.1|True
+3|4|False
+4|NULL|False
+NULL|4|False
+NULL|NULL|True"""
 
-    return spark.createDataFrame(mock_data)
-
-
-@pytest.fixture(scope="module", name="base_df2")
-def base_df2_fixture(spark):
-    mock_data = [
-        Row(
-            acct=10000001234,
-            dollar_amt=123,
-            super_duper_big_long_name="George Maharis",
-            float_fld=14530.1555,
-            date_fld=datetime.date(2017, 1, 1),
-        ),
-        Row(
-            acct=10000001235,
-            dollar_amt=0,
-            super_duper_big_long_name="Michael Bluth",
-            float_fld=1.0,
-            date_fld=datetime.date(2017, 1, 1),
-        ),
-        Row(
-            acct=10000001236,
-            dollar_amt=1345,
-            super_duper_big_long_name="George Bluth",
-            float_fld=None,
-            date_fld=datetime.date(2017, 1, 1),
-        ),
-        Row(
-            acct=10000001237,
-            dollar_amt=123456,
-            super_duper_big_long_name="Bob Loblaw",
-            float_fld=345.12,
-            date_fld=datetime.date(2017, 1, 1),
-        ),
-        Row(
-            acct=10000001239,
-            dollar_amt=1,
-            super_duper_big_long_name="Lucille Bluth",
-            float_fld=None,
-            date_fld=datetime.date(2017, 1, 1),
-        ),
-    ]
-
-    return spark.createDataFrame(mock_data)
-
-
-@pytest.fixture(scope="module", name="compare_df1")
-def compare_df1_fixture(spark):
-    mock_data2 = [
-        Row(
-            acct=10000001234,
-            dollar_amt=123.4,
-            name="George Michael Bluth",
-            float_fld=14530.155,
-            accnt_purge=False,
-        ),
-        Row(
-            acct=10000001235,
-            dollar_amt=0.45,
-            name="Michael Bluth",
-            float_fld=None,
-            accnt_purge=False,
-        ),
-        Row(
-            acct=10000001236,
-            dollar_amt=1345.0,
-            name="George Bluth",
-            float_fld=1.0,
-            accnt_purge=False,
-        ),
-        Row(
-            acct=10000001237,
-            dollar_amt=123456.0,
-            name="Bob Loblaw",
-            float_fld=345.12,
-            accnt_purge=False,
-        ),
-        Row(
-            acct=10000001238,
-            dollar_amt=1.05,
-            name="Loose Seal Bluth",
-            float_fld=111.0,
-            accnt_purge=True,
-        ),
-        Row(
-            acct=10000001238,
-            dollar_amt=1.05,
-            name="Loose Seal Bluth",
-            float_fld=111.0,
-            accnt_purge=True,
-        ),
-    ]
-
-    return spark.createDataFrame(mock_data2)
-
-
-@pytest.fixture(scope="module", name="compare_df2")
-def compare_df2_fixture(spark):
-    mock_data = [
-        Row(
-            acct=10000001234,
-            dollar_amt=123,
-            name="George Maharis",
-            float_fld=14530.1555,
-            date_fld=datetime.date(2017, 1, 1),
-        ),
-        Row(
-            acct=10000001235,
-            dollar_amt=0,
-            name="Michael Bluth",
-            float_fld=1.0,
-            date_fld=datetime.date(2017, 1, 1),
-        ),
-        Row(
-            acct=10000001236,
-            dollar_amt=1345,
-            name="George Bluth",
-            float_fld=None,
-            date_fld=datetime.date(2017, 1, 1),
-        ),
-        Row(
-            acct=10000001237,
-            dollar_amt=123456,
-            name="Bob Loblaw",
-            float_fld=345.12,
-            date_fld=datetime.date(2017, 1, 1),
-        ),
-        Row(
-            acct=10000001239,
-            dollar_amt=1,
-            name="Lucille Bluth",
-            float_fld=None,
-            date_fld=datetime.date(2017, 1, 1),
-        ),
-    ]
-
-    return spark.createDataFrame(mock_data)
-
-
-@pytest.fixture(scope="module", name="compare_df3")
-def compare_df3_fixture(spark):
-    mock_data2 = [
-        Row(
-            account_identifier=10000001234,
-            dollar_amount=123.4,
-            name="George Michael Bluth",
-            float_field=14530.155,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),
-        Row(
-            account_identifier=10000001235,
-            dollar_amount=0.45,
-            name="Michael Bluth",
-            float_field=1.0,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),
-        Row(
-            account_identifier=10000001236,
-            dollar_amount=1345.0,
-            name="George Bluth",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),
-        Row(
-            account_identifier=10000001237,
-            dollar_amount=123456.0,
-            name="Bob Loblaw",
-            float_field=345.12,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),
-        Row(
-            account_identifier=10000001239,
-            dollar_amount=1.05,
-            name="Lucille Bluth",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),
-    ]
-
-    return spark.createDataFrame(mock_data2)
-
-
-@pytest.fixture(scope="module", name="base_tol")
-def base_tol_fixture(spark):
-    tol_data1 = [
-        Row(
-            account_identifier=10000001234,
-            dollar_amount=123.4,
-            name="Franklin Delano Bluth",
-            float_field=14530.155,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),
-        Row(
-            account_identifier=10000001235,
-            dollar_amount=500.0,
-            name="Surely Funke",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),
-        Row(
-            account_identifier=10000001236,
-            dollar_amount=-1100.0,
-            name="Nichael Bluth",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),
-        Row(
-            account_identifier=10000001237,
-            dollar_amount=0.45,
-            name="Mr. F",
-            float_field=1.0,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),
-        Row(
-            account_identifier=10000001238,
-            dollar_amount=1345.0,
-            name="Steve Holt!",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),
-        Row(
-            account_identifier=10000001239,
-            dollar_amount=123456.0,
-            name="Blue Man Group",
-            float_field=345.12,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),
-        Row(
-            account_identifier=10000001240,
-            dollar_amount=1.1,
-            name="Her?",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),
-        Row(
-            account_identifier=10000001241,
-            dollar_amount=0.0,
-            name="Mrs. Featherbottom",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),
-        Row(
-            account_identifier=10000001242,
-            dollar_amount=0.0,
-            name="Ice",
-            float_field=345.12,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),
-        Row(
-            account_identifier=10000001243,
-            dollar_amount=-10.0,
-            name="Frank Wrench",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),
-        Row(
-            account_identifier=10000001244,
-            dollar_amount=None,
-            name="Lucille 2",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),
-        Row(
-            account_identifier=10000001245,
-            dollar_amount=0.009999,
-            name="Gene Parmesan",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),
-        Row(
-            account_identifier=10000001246,
-            dollar_amount=None,
-            name="Motherboy",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),
-    ]
-
-    return spark.createDataFrame(tol_data1)
-
-
-@pytest.fixture(scope="module", name="compare_abs_tol")
-def compare_tol2_fixture(spark):
-    tol_data2 = [
-        Row(
-            account_identifier=10000001234,
-            dollar_amount=123.4,
-            name="Franklin Delano Bluth",
-            float_field=14530.155,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),  # full match
-        Row(
-            account_identifier=10000001235,
-            dollar_amount=500.01,
-            name="Surely Funke",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # off by 0.01
-        Row(
-            account_identifier=10000001236,
-            dollar_amount=-1100.01,
-            name="Nichael Bluth",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # off by -0.01
-        Row(
-            account_identifier=10000001237,
-            dollar_amount=0.46000000001,
-            name="Mr. F",
-            float_field=1.0,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),  # off by 0.01000000001
-        Row(
-            account_identifier=10000001238,
-            dollar_amount=1344.8999999999,
-            name="Steve Holt!",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),  # off by -0.01000000001
-        Row(
-            account_identifier=10000001239,
-            dollar_amount=123456.0099999999,
-            name="Blue Man Group",
-            float_field=345.12,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),  # off by 0.00999999999
-        Row(
-            account_identifier=10000001240,
-            dollar_amount=1.090000001,
-            name="Her?",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # off by -0.00999999999
-        Row(
-            account_identifier=10000001241,
-            dollar_amount=0.0,
-            name="Mrs. Featherbottom",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # both zero
-        Row(
-            account_identifier=10000001242,
-            dollar_amount=1.0,
-            name="Ice",
-            float_field=345.12,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),  # base 0, compare 1
-        Row(
-            account_identifier=10000001243,
-            dollar_amount=0.0,
-            name="Frank Wrench",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # base -10, compare 0
-        Row(
-            account_identifier=10000001244,
-            dollar_amount=-1.0,
-            name="Lucille 2",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # base NULL, compare -1
-        Row(
-            account_identifier=10000001245,
-            dollar_amount=None,
-            name="Gene Parmesan",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # base 0.009999, compare NULL
-        Row(
-            account_identifier=10000001246,
-            dollar_amount=None,
-            name="Motherboy",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # both NULL
-    ]
-
-    return spark.createDataFrame(tol_data2)
-
-
-@pytest.fixture(scope="module", name="compare_rel_tol")
-def compare_tol3_fixture(spark):
-    tol_data3 = [
-        Row(
-            account_identifier=10000001234,
-            dollar_amount=123.4,
-            name="Franklin Delano Bluth",
-            float_field=14530.155,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),  # full match   #MATCH
-        Row(
-            account_identifier=10000001235,
-            dollar_amount=550.0,
-            name="Surely Funke",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # off by 10%   #MATCH
-        Row(
-            account_identifier=10000001236,
-            dollar_amount=-1000.0,
-            name="Nichael Bluth",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # off by -10%    #MATCH
-        Row(
-            account_identifier=10000001237,
-            dollar_amount=0.49501,
-            name="Mr. F",
-            float_field=1.0,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),  # off by greater than 10%
-        Row(
-            account_identifier=10000001238,
-            dollar_amount=1210.001,
-            name="Steve Holt!",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),  # off by greater than -10%
-        Row(
-            account_identifier=10000001239,
-            dollar_amount=135801.59999,
-            name="Blue Man Group",
-            float_field=345.12,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),  # off by just under 10%   #MATCH
-        Row(
-            account_identifier=10000001240,
-            dollar_amount=1.000001,
-            name="Her?",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # off by just under -10%   #MATCH
-        Row(
-            account_identifier=10000001241,
-            dollar_amount=0.0,
-            name="Mrs. Featherbottom",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # both zero   #MATCH
-        Row(
-            account_identifier=10000001242,
-            dollar_amount=1.0,
-            name="Ice",
-            float_field=345.12,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),  # base 0, compare 1
-        Row(
-            account_identifier=10000001243,
-            dollar_amount=0.0,
-            name="Frank Wrench",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # base -10, compare 0
-        Row(
-            account_identifier=10000001244,
-            dollar_amount=-1.0,
-            name="Lucille 2",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # base NULL, compare -1
-        Row(
-            account_identifier=10000001245,
-            dollar_amount=None,
-            name="Gene Parmesan",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # base 0.009999, compare NULL
-        Row(
-            account_identifier=10000001246,
-            dollar_amount=None,
-            name="Motherboy",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # both NULL  #MATCH
-    ]
-
-    return spark.createDataFrame(tol_data3)
-
-
-@pytest.fixture(scope="module", name="compare_both_tol")
-def compare_tol4_fixture(spark):
-    tol_data4 = [
-        Row(
-            account_identifier=10000001234,
-            dollar_amount=123.4,
-            name="Franklin Delano Bluth",
-            float_field=14530.155,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),  # full match
-        Row(
-            account_identifier=10000001235,
-            dollar_amount=550.01,
-            name="Surely Funke",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # off by 10% and +0.01
-        Row(
-            account_identifier=10000001236,
-            dollar_amount=-1000.01,
-            name="Nichael Bluth",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # off by -10% and -0.01
-        Row(
-            account_identifier=10000001237,
-            dollar_amount=0.505000000001,
-            name="Mr. F",
-            float_field=1.0,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),  # off by greater than 10% and +0.01
-        Row(
-            account_identifier=10000001238,
-            dollar_amount=1209.98999,
-            name="Steve Holt!",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),  # off by greater than -10% and -0.01
-        Row(
-            account_identifier=10000001239,
-            dollar_amount=135801.609999,
-            name="Blue Man Group",
-            float_field=345.12,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),  # off by just under 10% and just under +0.01
-        Row(
-            account_identifier=10000001240,
-            dollar_amount=0.99000001,
-            name="Her?",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # off by just under -10% and just under -0.01
-        Row(
-            account_identifier=10000001241,
-            dollar_amount=0.0,
-            name="Mrs. Featherbottom",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # both zero
-        Row(
-            account_identifier=10000001242,
-            dollar_amount=1.0,
-            name="Ice",
-            float_field=345.12,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=False,
-        ),  # base 0, compare 1
-        Row(
-            account_identifier=10000001243,
-            dollar_amount=0.0,
-            name="Frank Wrench",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # base -10, compare 0
-        Row(
-            account_identifier=10000001244,
-            dollar_amount=-1.0,
-            name="Lucille 2",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # base NULL, compare -1
-        Row(
-            account_identifier=10000001245,
-            dollar_amount=None,
-            name="Gene Parmesan",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # base 0.009999, compare NULL
-        Row(
-            account_identifier=10000001246,
-            dollar_amount=None,
-            name="Motherboy",
-            float_field=None,
-            date_field=datetime.date(2017, 1, 1),
-            accnt_purge=True,
-        ),  # both NULL
-    ]
-
-    return spark.createDataFrame(tol_data4)
-
-
-@pytest.fixture(scope="module", name="base_td")
-def base_td_fixture(spark):
-    mock_data = [
-        Row(
-            acct=10000001234,
-            acct_seq=0,
-            stat_cd="*2",
-            open_dt=datetime.date(2017, 5, 1),
-            cd="0001",
-        ),
-        Row(
-            acct=10000001235,
-            acct_seq=0,
-            stat_cd="V1",
-            open_dt=datetime.date(2017, 5, 2),
-            cd="0002",
-        ),
-        Row(
-            acct=10000001236,
-            acct_seq=0,
-            stat_cd="V2",
-            open_dt=datetime.date(2017, 5, 3),
-            cd="0003",
-        ),
-        Row(
-            acct=10000001237,
-            acct_seq=0,
-            stat_cd="*2",
-            open_dt=datetime.date(2017, 5, 4),
-            cd="0004",
-        ),
-        Row(
-            acct=10000001238,
-            acct_seq=0,
-            stat_cd="*2",
-            open_dt=datetime.date(2017, 5, 5),
-            cd="0005",
-        ),
-    ]
-
-    return spark.createDataFrame(mock_data)
-
-
-@pytest.fixture(scope="module", name="compare_source")
-def compare_source_fixture(spark):
-    mock_data = [
-        Row(
-            ACCOUNT_IDENTIFIER=10000001234,
-            SEQ_NUMBER=0,
-            STATC=None,
-            ACCOUNT_OPEN=2017121,
-            CODE=1.0,
-        ),
-        Row(
-            ACCOUNT_IDENTIFIER=10000001235,
-            SEQ_NUMBER=0,
-            STATC="V1",
-            ACCOUNT_OPEN=2017122,
-            CODE=2.0,
-        ),
-        Row(
-            ACCOUNT_IDENTIFIER=10000001236,
-            SEQ_NUMBER=0,
-            STATC="V2",
-            ACCOUNT_OPEN=2017123,
-            CODE=3.0,
-        ),
-        Row(
-            ACCOUNT_IDENTIFIER=10000001237,
-            SEQ_NUMBER=0,
-            STATC="V3",
-            ACCOUNT_OPEN=2017124,
-            CODE=4.0,
-        ),
-        Row(
-            ACCOUNT_IDENTIFIER=10000001238,
-            SEQ_NUMBER=0,
-            STATC=None,
-            ACCOUNT_OPEN=2017125,
-            CODE=5.0,
-        ),
-    ]
-
-    return spark.createDataFrame(mock_data)
-
-
-@pytest.fixture(scope="module", name="base_decimal")
-def base_decimal_fixture(spark):
-    mock_data = [
-        Row(acct=10000001234, dollar_amt=Decimal(123.4)),
-        Row(acct=10000001235, dollar_amt=Decimal(0.45)),
-    ]
-
-    return spark.createDataFrame(
-        mock_data,
-        schema=StructType(
-            [
-                StructField("acct", LongType(), True),
-                StructField("dollar_amt", DecimalType(8, 2), True),
-            ]
-        ),
+    df = ps.from_pandas(pd.read_csv(StringIO(data), sep="|"))
+    actual_out = columns_equal(df.a, df.b, abs_tol=0.2)
+    expect_out = df["expected"]
+    assert_series_equal(
+        expect_out.to_pandas(), actual_out.to_pandas(), check_names=False
     )
 
 
-@pytest.fixture(scope="module", name="compare_decimal")
-def compare_decimal_fixture(spark):
-    mock_data = [
-        Row(acct=10000001234, dollar_amt=123.4),
-        Row(acct=10000001235, dollar_amt=0.456),
-    ]
-
-    return spark.createDataFrame(mock_data)
-
-
-@pytest.fixture(scope="module", name="comparison_abs_tol")
-def comparison_abs_tol_fixture(base_tol, compare_abs_tol, spark):
-    return SparkCompare(
-        spark,
-        base_tol,
-        compare_abs_tol,
-        join_columns=["account_identifier"],
-        abs_tol=0.01,
+@pandas_version
+def test_numeric_columns_equal_rel():
+    data = """a|b|expected
+1|1|True
+2|2.1|True
+3|4|False
+4|NULL|False
+NULL|4|False
+NULL|NULL|True"""
+    df = ps.from_pandas(pd.read_csv(StringIO(data), sep="|"))
+    actual_out = columns_equal(df.a, df.b, rel_tol=0.2)
+    expect_out = df["expected"]
+    assert_series_equal(
+        expect_out.to_pandas(), actual_out.to_pandas(), check_names=False
     )
 
 
-@pytest.fixture(scope="module", name="comparison_rel_tol")
-def comparison_rel_tol_fixture(base_tol, compare_rel_tol, spark):
-    return SparkCompare(
-        spark,
-        base_tol,
-        compare_rel_tol,
-        join_columns=["account_identifier"],
-        rel_tol=0.1,
+@pandas_version
+def test_string_columns_equal():
+    data = """a|b|expected
+Hi|Hi|True
+Yo|Yo|True
+Hey|Hey |False
+résumé|resume|False
+résumé|résumé|True
+💩|💩|True
+💩|🤔|False
+ | |True
+  | |False
+datacompy|DataComPy|False
+something||False
+|something|False
+||True"""
+    df = ps.from_pandas(pd.read_csv(StringIO(data), sep="|"))
+    actual_out = columns_equal(df.a, df.b, rel_tol=0.2)
+    expect_out = df["expected"]
+    assert_series_equal(
+        expect_out.to_pandas(), actual_out.to_pandas(), check_names=False
     )
 
 
-@pytest.fixture(scope="module", name="comparison_both_tol")
-def comparison_both_tol_fixture(base_tol, compare_both_tol, spark):
-    return SparkCompare(
-        spark,
-        base_tol,
-        compare_both_tol,
-        join_columns=["account_identifier"],
-        rel_tol=0.1,
-        abs_tol=0.01,
+@pandas_version
+def test_string_columns_equal_with_ignore_spaces():
+    data = """a|b|expected
+Hi|Hi|True
+Yo|Yo|True
+Hey|Hey |True
+résumé|resume|False
+résumé|résumé|True
+💩|💩|True
+💩|🤔|False
+ | |True
+  |       |True
+datacompy|DataComPy|False
+something||False
+|something|False
+||True"""
+    df = ps.from_pandas(pd.read_csv(StringIO(data), sep="|"))
+    actual_out = columns_equal(df.a, df.b, rel_tol=0.2, ignore_spaces=True)
+    expect_out = df["expected"]
+    assert_series_equal(
+        expect_out.to_pandas(), actual_out.to_pandas(), check_names=False
     )
 
 
-@pytest.fixture(scope="module", name="comparison_neg_tol")
-def comparison_neg_tol_fixture(base_tol, compare_both_tol, spark):
-    return SparkCompare(
-        spark,
-        base_tol,
-        compare_both_tol,
-        join_columns=["account_identifier"],
-        rel_tol=-0.2,
-        abs_tol=0.01,
+@pandas_version
+def test_string_columns_equal_with_ignore_spaces_and_case():
+    data = """a|b|expected
+Hi|Hi|True
+Yo|Yo|True
+Hey|Hey |True
+résumé|resume|False
+résumé|résumé|True
+💩|💩|True
+💩|🤔|False
+ | |True
+  |       |True
+datacompy|DataComPy|True
+something||False
+|something|False
+||True"""
+    df = ps.from_pandas(pd.read_csv(StringIO(data), sep="|"))
+    actual_out = columns_equal(
+        df.a, df.b, rel_tol=0.2, ignore_spaces=True, ignore_case=True
+    )
+    expect_out = df["expected"]
+    assert_series_equal(
+        expect_out.to_pandas(), actual_out.to_pandas(), check_names=False
     )
 
 
-@pytest.fixture(scope="module", name="show_all_columns_and_match_rate")
-def show_all_columns_and_match_rate_fixture(base_tol, compare_both_tol, spark):
-    return SparkCompare(
-        spark,
-        base_tol,
-        compare_both_tol,
-        join_columns=["account_identifier"],
-        show_all_columns=True,
-        match_rates=True,
+@pandas_version
+def test_date_columns_equal(tmp_path):
+    data = """a|b|expected
+2017-01-01|2017-01-01|True
+2017-01-02|2017-01-02|True
+2017-10-01|2017-10-10|False
+2017-01-01||False
+|2017-01-01|False
+||True"""
+    df = ps.from_pandas(pd.read_csv(StringIO(data), sep="|"))
+    # First compare just the strings
+    actual_out = columns_equal(df.a, df.b, rel_tol=0.2)
+    expect_out = df["expected"]
+    assert_series_equal(
+        expect_out.to_pandas(), actual_out.to_pandas(), check_names=False
+    )
+
+    # Then compare converted to datetime objects
+    df["a"] = ps.to_datetime(df["a"])
+    df["b"] = ps.to_datetime(df["b"])
+    actual_out = columns_equal(df.a, df.b, rel_tol=0.2)
+    expect_out = df["expected"]
+    assert_series_equal(
+        expect_out.to_pandas(), actual_out.to_pandas(), check_names=False
+    )
+    # and reverse
+    actual_out_rev = columns_equal(df.b, df.a, rel_tol=0.2)
+    assert_series_equal(
+        expect_out.to_pandas(), actual_out_rev.to_pandas(), check_names=False
     )
 
 
-@pytest.fixture(scope="module", name="comparison_kd1")
-def comparison_known_diffs1(base_td, compare_source, spark):
-    return SparkCompare(
-        spark,
-        base_td,
-        compare_source,
-        join_columns=[("acct", "ACCOUNT_IDENTIFIER"), ("acct_seq", "SEQ_NUMBER")],
-        column_mapping=[
-            ("stat_cd", "STATC"),
-            ("open_dt", "ACCOUNT_OPEN"),
-            ("cd", "CODE"),
-        ],
-        known_differences=[
+@pandas_version
+def test_date_columns_equal_with_ignore_spaces(tmp_path):
+    data = """a|b|expected
+2017-01-01|2017-01-01   |True
+2017-01-02  |2017-01-02|True
+2017-10-01  |2017-10-10   |False
+2017-01-01||False
+|2017-01-01|False
+||True"""
+    df = ps.from_pandas(pd.read_csv(StringIO(data), sep="|"))
+    # First compare just the strings
+    actual_out = columns_equal(df.a, df.b, rel_tol=0.2, ignore_spaces=True)
+    expect_out = df["expected"]
+    assert_series_equal(
+        expect_out.to_pandas(), actual_out.to_pandas(), check_names=False
+    )
+
+    # Then compare converted to datetime objects
+    df["a"] = ps.to_datetime(df["a"], errors="coerce")
+    df["b"] = ps.to_datetime(df["b"], errors="coerce")
+    actual_out = columns_equal(df.a, df.b, rel_tol=0.2, ignore_spaces=True)
+    expect_out = df["expected"]
+    assert_series_equal(
+        expect_out.to_pandas(), actual_out.to_pandas(), check_names=False
+    )
+    # and reverse
+    actual_out_rev = columns_equal(df.b, df.a, rel_tol=0.2, ignore_spaces=True)
+    assert_series_equal(
+        expect_out.to_pandas(), actual_out_rev.to_pandas(), check_names=False
+    )
+
+
+@pandas_version
+def test_date_columns_equal_with_ignore_spaces_and_case(tmp_path):
+    data = """a|b|expected
+2017-01-01|2017-01-01   |True
+2017-01-02  |2017-01-02|True
+2017-10-01  |2017-10-10   |False
+2017-01-01||False
+|2017-01-01|False
+||True"""
+    df = ps.from_pandas(pd.read_csv(StringIO(data), sep="|"))
+    # First compare just the strings
+    actual_out = columns_equal(
+        df.a, df.b, rel_tol=0.2, ignore_spaces=True, ignore_case=True
+    )
+    expect_out = df["expected"]
+    assert_series_equal(
+        expect_out.to_pandas(), actual_out.to_pandas(), check_names=False
+    )
+
+    # Then compare converted to datetime objects
+    df["a"] = ps.to_datetime(df["a"], errors="coerce")
+    df["b"] = ps.to_datetime(df["b"], errors="coerce")
+    actual_out = columns_equal(df.a, df.b, rel_tol=0.2, ignore_spaces=True)
+    expect_out = df["expected"]
+    assert_series_equal(
+        expect_out.to_pandas(), actual_out.to_pandas(), check_names=False
+    )
+    # and reverse
+    actual_out_rev = columns_equal(df.b, df.a, rel_tol=0.2, ignore_spaces=True)
+    assert_series_equal(
+        expect_out.to_pandas(), actual_out_rev.to_pandas(), check_names=False
+    )
+
+
+@pandas_version
+def test_date_columns_unequal():
+    """I want datetime fields to match with dates stored as strings"""
+    df = ps.DataFrame([{"a": "2017-01-01", "b": "2017-01-02"}, {"a": "2017-01-01"}])
+    df["a_dt"] = ps.to_datetime(df["a"])
+    df["b_dt"] = ps.to_datetime(df["b"])
+    assert columns_equal(df.a, df.a_dt).all()
+    assert columns_equal(df.b, df.b_dt).all()
+    assert columns_equal(df.a_dt, df.a).all()
+    assert columns_equal(df.b_dt, df.b).all()
+    assert not columns_equal(df.b_dt, df.a).any()
+    assert not columns_equal(df.a_dt, df.b).any()
+    assert not columns_equal(df.a, df.b_dt).any()
+    assert not columns_equal(df.b, df.a_dt).any()
+
+
+@pandas_version
+def test_bad_date_columns():
+    """If strings can't be coerced into dates then it should be false for the
+    whole column.
+    """
+    df = ps.DataFrame(
+        [{"a": "2017-01-01", "b": "2017-01-01"}, {"a": "2017-01-01", "b": "217-01-01"}]
+    )
+    df["a_dt"] = ps.to_datetime(df["a"])
+    assert not columns_equal(df.a_dt, df.b).any()
+
+
+@pandas_version
+def test_rounded_date_columns():
+    """If strings can't be coerced into dates then it should be false for the
+    whole column.
+    """
+    df = ps.DataFrame(
+        [
+            {"a": "2017-01-01", "b": "2017-01-01 00:00:00.000000", "exp": True},
+            {"a": "2017-01-01", "b": "2017-01-01 00:00:00.123456", "exp": False},
+            {"a": "2017-01-01", "b": "2017-01-01 00:00:01.000000", "exp": False},
+            {"a": "2017-01-01", "b": "2017-01-01 00:00:00", "exp": True},
+        ]
+    )
+    df["a_dt"] = ps.to_datetime(df["a"])
+    actual = columns_equal(df.a_dt, df.b)
+    expected = df["exp"]
+    assert_series_equal(actual.to_pandas(), expected.to_pandas(), check_names=False)
+
+
+@pandas_version
+def test_decimal_float_columns_equal():
+    df = ps.DataFrame(
+        [
+            {"a": Decimal("1"), "b": 1, "expected": True},
+            {"a": Decimal("1.3"), "b": 1.3, "expected": True},
+            {"a": Decimal("1.000003"), "b": 1.000003, "expected": True},
+            {"a": Decimal("1.000000004"), "b": 1.000000003, "expected": False},
+            {"a": Decimal("1.3"), "b": 1.2, "expected": False},
+            {"a": np.nan, "b": np.nan, "expected": True},
+            {"a": np.nan, "b": 1, "expected": False},
+            {"a": Decimal("1"), "b": np.nan, "expected": False},
+        ]
+    )
+    actual_out = columns_equal(df.a, df.b)
+    expect_out = df["expected"]
+    assert_series_equal(
+        expect_out.to_pandas(), actual_out.to_pandas(), check_names=False
+    )
+
+
+@pandas_version
+def test_decimal_float_columns_equal_rel():
+    df = ps.DataFrame(
+        [
+            {"a": Decimal("1"), "b": 1, "expected": True},
+            {"a": Decimal("1.3"), "b": 1.3, "expected": True},
+            {"a": Decimal("1.000003"), "b": 1.000003, "expected": True},
+            {"a": Decimal("1.000000004"), "b": 1.000000003, "expected": True},
+            {"a": Decimal("1.3"), "b": 1.2, "expected": False},
+            {"a": np.nan, "b": np.nan, "expected": True},
+            {"a": np.nan, "b": 1, "expected": False},
+            {"a": Decimal("1"), "b": np.nan, "expected": False},
+        ]
+    )
+    actual_out = columns_equal(df.a, df.b, abs_tol=0.001)
+    expect_out = df["expected"]
+    assert_series_equal(
+        expect_out.to_pandas(), actual_out.to_pandas(), check_names=False
+    )
+
+
+@pandas_version
+def test_decimal_columns_equal():
+    df = ps.DataFrame(
+        [
+            {"a": Decimal("1"), "b": Decimal("1"), "expected": True},
+            {"a": Decimal("1.3"), "b": Decimal("1.3"), "expected": True},
+            {"a": Decimal("1.000003"), "b": Decimal("1.000003"), "expected": True},
             {
-                "name": "Left-padded, four-digit numeric code",
-                "types": datacompy.NUMERIC_SPARK_TYPES,
-                "transformation": "lpad(cast({input} AS bigint), 4, '0')",
+                "a": Decimal("1.000000004"),
+                "b": Decimal("1.000000003"),
+                "expected": False,
             },
+            {"a": Decimal("1.3"), "b": Decimal("1.2"), "expected": False},
+            {"a": np.nan, "b": np.nan, "expected": True},
+            {"a": np.nan, "b": Decimal("1"), "expected": False},
+            {"a": Decimal("1"), "b": np.nan, "expected": False},
+        ]
+    )
+    actual_out = columns_equal(df.a, df.b)
+    expect_out = df["expected"]
+    assert_series_equal(
+        expect_out.to_pandas(), actual_out.to_pandas(), check_names=False
+    )
+
+
+@pandas_version
+def test_decimal_columns_equal_rel():
+    df = ps.DataFrame(
+        [
+            {"a": Decimal("1"), "b": Decimal("1"), "expected": True},
+            {"a": Decimal("1.3"), "b": Decimal("1.3"), "expected": True},
+            {"a": Decimal("1.000003"), "b": Decimal("1.000003"), "expected": True},
             {
-                "name": "Null to *2",
-                "types": ["string"],
-                "transformation": "case when {input} is null then '*2' else {input} end",
+                "a": Decimal("1.000000004"),
+                "b": Decimal("1.000000003"),
+                "expected": True,
             },
-            {
-                "name": "Julian date -> date",
-                "types": ["bigint"],
-                "transformation": "to_date(cast(unix_timestamp(cast({input} AS string), 'yyyyDDD') AS timestamp))",
-            },
-        ],
-    )
-
-
-@pytest.fixture(scope="module", name="comparison_kd2")
-def comparison_known_diffs2(base_td, compare_source, spark):
-    return SparkCompare(
-        spark,
-        base_td,
-        compare_source,
-        join_columns=[("acct", "ACCOUNT_IDENTIFIER"), ("acct_seq", "SEQ_NUMBER")],
-        column_mapping=[
-            ("stat_cd", "STATC"),
-            ("open_dt", "ACCOUNT_OPEN"),
-            ("cd", "CODE"),
-        ],
-        known_differences=[
-            {
-                "name": "Left-padded, four-digit numeric code",
-                "types": datacompy.NUMERIC_SPARK_TYPES,
-                "transformation": "lpad(cast({input} AS bigint), 4, '0')",
-            },
-            {
-                "name": "Null to *2",
-                "types": ["string"],
-                "transformation": "case when {input} is null then '*2' else {input} end",
-            },
-        ],
-    )
-
-
-@pytest.fixture(scope="module", name="comparison1")
-def comparison1_fixture(base_df1, compare_df1, spark):
-    return SparkCompare(
-        spark,
-        base_df1,
-        compare_df1,
-        join_columns=["acct"],
-        cache_intermediates=CACHE_INTERMEDIATES,
-    )
-
-
-@pytest.fixture(scope="module", name="comparison2")
-def comparison2_fixture(base_df1, compare_df2, spark):
-    return SparkCompare(spark, base_df1, compare_df2, join_columns=["acct"])
-
-
-@pytest.fixture(scope="module", name="comparison3")
-def comparison3_fixture(base_df1, compare_df3, spark):
-    return SparkCompare(
-        spark,
-        base_df1,
-        compare_df3,
-        join_columns=[("acct", "account_identifier")],
-        column_mapping=[
-            ("dollar_amt", "dollar_amount"),
-            ("float_fld", "float_field"),
-            ("date_fld", "date_field"),
-        ],
-        cache_intermediates=CACHE_INTERMEDIATES,
-    )
-
-
-@pytest.fixture(scope="module", name="comparison4")
-def comparison4_fixture(base_df2, compare_df1, spark):
-    return SparkCompare(
-        spark,
-        base_df2,
-        compare_df1,
-        join_columns=["acct"],
-        column_mapping=[("super_duper_big_long_name", "name")],
-    )
-
-
-@pytest.fixture(scope="module", name="comparison_decimal")
-def comparison_decimal_fixture(base_decimal, compare_decimal, spark):
-    return SparkCompare(spark, base_decimal, compare_decimal, join_columns=["acct"])
-
-
-def test_absolute_tolerances(comparison_abs_tol):
-    stdout = io.StringIO()
-
-    comparison_abs_tol.report(file=stdout)
-    stdout.seek(0)
-    assert "****** Row Comparison ******" in stdout.getvalue()
-    assert "Number of rows with some columns unequal: 6" in stdout.getvalue()
-    assert "Number of rows with all columns equal: 7" in stdout.getvalue()
-    assert "Number of columns compared with some values unequal: 1" in stdout.getvalue()
-    assert "Number of columns compared with all values equal: 4" in stdout.getvalue()
-
-
-def test_relative_tolerances(comparison_rel_tol):
-    stdout = io.StringIO()
-
-    comparison_rel_tol.report(file=stdout)
-    stdout.seek(0)
-    assert "****** Row Comparison ******" in stdout.getvalue()
-    assert "Number of rows with some columns unequal: 6" in stdout.getvalue()
-    assert "Number of rows with all columns equal: 7" in stdout.getvalue()
-    assert "Number of columns compared with some values unequal: 1" in stdout.getvalue()
-    assert "Number of columns compared with all values equal: 4" in stdout.getvalue()
-
-
-def test_both_tolerances(comparison_both_tol):
-    stdout = io.StringIO()
-
-    comparison_both_tol.report(file=stdout)
-    stdout.seek(0)
-    assert "****** Row Comparison ******" in stdout.getvalue()
-    assert "Number of rows with some columns unequal: 6" in stdout.getvalue()
-    assert "Number of rows with all columns equal: 7" in stdout.getvalue()
-    assert "Number of columns compared with some values unequal: 1" in stdout.getvalue()
-    assert "Number of columns compared with all values equal: 4" in stdout.getvalue()
-
-
-def test_negative_tolerances(spark, base_tol, compare_both_tol):
-    with pytest.raises(ValueError, match="Please enter positive valued tolerances"):
-        comp = SparkCompare(
-            spark,
-            base_tol,
-            compare_both_tol,
-            join_columns=["account_identifier"],
-            rel_tol=-0.2,
-            abs_tol=0.01,
-        )
-        comp.report()
-        pass
-
-
-def test_show_all_columns_and_match_rate(show_all_columns_and_match_rate):
-    stdout = io.StringIO()
-
-    show_all_columns_and_match_rate.report(file=stdout)
-
-    assert "****** Columns with Equal/Unequal Values ******" in stdout.getvalue()
-    assert (
-        "accnt_purge       accnt_purge          boolean        boolean               13             0     100.00000"
-        in stdout.getvalue()
-    )
-    assert (
-        "date_field        date_field           date           date                  13             0     100.00000"
-        in stdout.getvalue()
-    )
-    assert (
-        "dollar_amount     dollar_amount        double         double                 3            10      23.07692"
-        in stdout.getvalue()
-    )
-    assert (
-        "float_field       float_field          double         double                13             0     100.00000"
-        in stdout.getvalue()
-    )
-    assert (
-        "name              name                 string         string                13             0     100.00000"
-        in stdout.getvalue()
-    )
-
-
-def test_decimal_comparisons():
-    true_decimals = ["decimal", "decimal()", "decimal(20, 10)"]
-    assert all(v in datacompy.NUMERIC_SPARK_TYPES for v in true_decimals)
-
-
-def test_decimal_comparator_acts_like_string():
-    acc = False
-    for t in datacompy.NUMERIC_SPARK_TYPES:
-        acc = acc or (len(t) > 2 and t[0:3] == "dec")
-    assert acc
-
-
-def test_decimals_and_doubles_are_comparable():
-    assert _is_comparable("double", "decimal(10, 2)")
-
-
-def test_report_outputs_the_column_summary(comparison1):
-    stdout = io.StringIO()
-
-    comparison1.report(file=stdout)
-
-    assert "****** Column Summary ******" in stdout.getvalue()
-    assert "Number of columns in common with matching schemas: 3" in stdout.getvalue()
-    assert "Number of columns in common with schema differences: 1" in stdout.getvalue()
-    assert "Number of columns in base but not compare: 1" in stdout.getvalue()
-    assert "Number of columns in compare but not base: 1" in stdout.getvalue()
-
-
-def test_report_outputs_the_column_summary_for_identical_schemas(comparison2):
-    stdout = io.StringIO()
-
-    comparison2.report(file=stdout)
-
-    assert "****** Column Summary ******" in stdout.getvalue()
-    assert "Number of columns in common with matching schemas: 5" in stdout.getvalue()
-    assert "Number of columns in common with schema differences: 0" in stdout.getvalue()
-    assert "Number of columns in base but not compare: 0" in stdout.getvalue()
-    assert "Number of columns in compare but not base: 0" in stdout.getvalue()
-
-
-def test_report_outputs_the_column_summary_for_differently_named_columns(comparison3):
-    stdout = io.StringIO()
-
-    comparison3.report(file=stdout)
-
-    assert "****** Column Summary ******" in stdout.getvalue()
-    assert "Number of columns in common with matching schemas: 4" in stdout.getvalue()
-    assert "Number of columns in common with schema differences: 1" in stdout.getvalue()
-    assert "Number of columns in base but not compare: 0" in stdout.getvalue()
-    assert "Number of columns in compare but not base: 1" in stdout.getvalue()
-
-
-def test_report_outputs_the_row_summary(comparison1):
-    stdout = io.StringIO()
-
-    comparison1.report(file=stdout)
-
-    assert "****** Row Summary ******" in stdout.getvalue()
-    assert "Number of rows in common: 4" in stdout.getvalue()
-    assert "Number of rows in base but not compare: 1" in stdout.getvalue()
-    assert "Number of rows in compare but not base: 1" in stdout.getvalue()
-    assert "Number of duplicate rows found in base: 0" in stdout.getvalue()
-    assert "Number of duplicate rows found in compare: 1" in stdout.getvalue()
-
-
-def test_report_outputs_the_row_equality_comparison(comparison1):
-    stdout = io.StringIO()
-
-    comparison1.report(file=stdout)
-
-    assert "****** Row Comparison ******" in stdout.getvalue()
-    assert "Number of rows with some columns unequal: 3" in stdout.getvalue()
-    assert "Number of rows with all columns equal: 1" in stdout.getvalue()
-
-
-def test_report_outputs_the_row_summary_for_differently_named_columns(comparison3):
-    stdout = io.StringIO()
-
-    comparison3.report(file=stdout)
-
-    assert "****** Row Summary ******" in stdout.getvalue()
-    assert "Number of rows in common: 5" in stdout.getvalue()
-    assert "Number of rows in base but not compare: 0" in stdout.getvalue()
-    assert "Number of rows in compare but not base: 0" in stdout.getvalue()
-    assert "Number of duplicate rows found in base: 0" in stdout.getvalue()
-    assert "Number of duplicate rows found in compare: 0" in stdout.getvalue()
-
-
-def test_report_outputs_the_row_equality_comparison_for_differently_named_columns(
-    comparison3,
-):
-    stdout = io.StringIO()
-
-    comparison3.report(file=stdout)
-
-    assert "****** Row Comparison ******" in stdout.getvalue()
-    assert "Number of rows with some columns unequal: 3" in stdout.getvalue()
-    assert "Number of rows with all columns equal: 2" in stdout.getvalue()
-
-
-def test_report_outputs_column_detail_for_columns_in_only_one_dataframe(comparison1):
-    stdout = io.StringIO()
-
-    comparison1.report(file=stdout)
-    comparison1.report()
-    assert "****** Columns In Base Only ******" in stdout.getvalue()
-    r2 = r"""Column\s*Name \s* Dtype \n -+ \s+ -+ \ndate_fld \s+ date"""
-    assert re.search(r2, str(stdout.getvalue()), re.X) is not None
-
-
-def test_report_outputs_column_detail_for_columns_in_only_compare_dataframe(
-    comparison1,
-):
-    stdout = io.StringIO()
-
-    comparison1.report(file=stdout)
-    comparison1.report()
-    assert "****** Columns In Compare Only ******" in stdout.getvalue()
-    r2 = r"""Column\s*Name \s* Dtype \n -+ \s+ -+ \n accnt_purge \s+  boolean"""
-    assert re.search(r2, str(stdout.getvalue()), re.X) is not None
-
-
-def test_report_outputs_schema_difference_details(comparison1):
-    stdout = io.StringIO()
-
-    comparison1.report(file=stdout)
-
-    assert "****** Schema Differences ******" in stdout.getvalue()
-    assert re.search(
-        r"""Base\sColumn\sName \s+ Compare\sColumn\sName \s+ Base\sDtype \s+ Compare\sDtype \n
-            -+ \s+ -+ \s+ -+ \s+ -+ \n
-            dollar_amt \s+ dollar_amt \s+ bigint \s+ double""",
-        stdout.getvalue(),
-        re.X,
-    )
-
-
-def test_report_outputs_schema_difference_details_for_differently_named_columns(
-    comparison3,
-):
-    stdout = io.StringIO()
-
-    comparison3.report(file=stdout)
-
-    assert "****** Schema Differences ******" in stdout.getvalue()
-    assert re.search(
-        r"""Base\sColumn\sName \s+ Compare\sColumn\sName \s+ Base\sDtype \s+ Compare\sDtype \n
-            -+ \s+ -+ \s+ -+ \s+ -+ \n
-            dollar_amt \s+ dollar_amount \s+ bigint \s+ double""",
-        stdout.getvalue(),
-        re.X,
-    )
-
-
-def test_column_comparison_outputs_number_of_columns_with_differences(comparison1):
-    stdout = io.StringIO()
-
-    comparison1.report(file=stdout)
-
-    assert "****** Column Comparison ******" in stdout.getvalue()
-    assert "Number of columns compared with some values unequal: 3" in stdout.getvalue()
-    assert "Number of columns compared with all values equal: 0" in stdout.getvalue()
-
-
-def test_column_comparison_outputs_all_columns_equal_for_identical_dataframes(
-    comparison2,
-):
-    stdout = io.StringIO()
-
-    comparison2.report(file=stdout)
-
-    assert "****** Column Comparison ******" in stdout.getvalue()
-    assert "Number of columns compared with some values unequal: 0" in stdout.getvalue()
-    assert "Number of columns compared with all values equal: 4" in stdout.getvalue()
-
-
-def test_column_comparison_outputs_number_of_columns_with_differences_for_differently_named_columns(
-    comparison3,
-):
-    stdout = io.StringIO()
-
-    comparison3.report(file=stdout)
-
-    assert "****** Column Comparison ******" in stdout.getvalue()
-    assert "Number of columns compared with some values unequal: 3" in stdout.getvalue()
-    assert "Number of columns compared with all values equal: 1" in stdout.getvalue()
-
-
-def test_column_comparison_outputs_number_of_columns_with_differences_for_known_diffs(
-    comparison_kd1,
-):
-    stdout = io.StringIO()
-
-    comparison_kd1.report(file=stdout)
-
-    assert "****** Column Comparison ******" in stdout.getvalue()
-    assert (
-        "Number of columns compared with unexpected differences in some values: 1"
-        in stdout.getvalue()
-    )
-    assert (
-        "Number of columns compared with all values equal but known differences found: 2"
-        in stdout.getvalue()
-    )
-    assert (
-        "Number of columns compared with all values completely equal: 0"
-        in stdout.getvalue()
-    )
-
-
-def test_column_comparison_outputs_number_of_columns_with_differences_for_custom_known_diffs(
-    comparison_kd2,
-):
-    stdout = io.StringIO()
-
-    comparison_kd2.report(file=stdout)
-
-    assert "****** Column Comparison ******" in stdout.getvalue()
-    assert (
-        "Number of columns compared with unexpected differences in some values: 2"
-        in stdout.getvalue()
-    )
-    assert (
-        "Number of columns compared with all values equal but known differences found: 1"
-        in stdout.getvalue()
-    )
-    assert (
-        "Number of columns compared with all values completely equal: 0"
-        in stdout.getvalue()
-    )
-
-
-def test_columns_with_unequal_values_show_mismatch_counts(comparison1):
-    stdout = io.StringIO()
-
-    comparison1.report(file=stdout)
-
-    assert "****** Columns with Unequal Values ******" in stdout.getvalue()
-    assert re.search(
-        r"""Base\s*Column\s*Name \s+ Compare\s*Column\s*Name \s+ Base\s*Dtype \s+ Compare\sDtype \s*
-            \#\sMatches \s* \#\sMismatches \n
-            -+ \s+ -+ \s+ -+ \s+ -+ \s+ -+ \s+ -+""",
-        stdout.getvalue(),
-        re.X,
-    )
-    assert re.search(
-        r"""dollar_amt \s+ dollar_amt \s+ bigint \s+ double \s+ 2 \s+ 2""",
-        stdout.getvalue(),
-        re.X,
-    )
-    assert re.search(
-        r"""float_fld \s+ float_fld \s+ double \s+ double \s+ 1 \s+ 3""",
-        stdout.getvalue(),
-        re.X,
-    )
-    assert re.search(
-        r"""name \s+ name \s+ string \s+ string \s+ 3 \s+ 1""", stdout.getvalue(), re.X
-    )
-
-
-def test_columns_with_different_names_with_unequal_values_show_mismatch_counts(
-    comparison3,
-):
-    stdout = io.StringIO()
-
-    comparison3.report(file=stdout)
-
-    assert "****** Columns with Unequal Values ******" in stdout.getvalue()
-    assert re.search(
-        r"""Base\s*Column\s*Name \s+ Compare\s*Column\s*Name \s+ Base\s*Dtype \s+ Compare\sDtype \s*
-            \#\sMatches \s* \#\sMismatches \n
-            -+ \s+ -+ \s+ -+ \s+ -+ \s+ -+ \s+ -+""",
-        stdout.getvalue(),
-        re.X,
-    )
-    assert re.search(
-        r"""dollar_amt \s+ dollar_amount \s+ bigint \s+ double \s+ 2 \s+ 3""",
-        stdout.getvalue(),
-        re.X,
-    )
-    assert re.search(
-        r"""float_fld \s+ float_field \s+ double \s+ double \s+ 4 \s+ 1""",
-        stdout.getvalue(),
-        re.X,
-    )
-    assert re.search(
-        r"""name \s+ name \s+ string \s+ string \s+ 4 \s+ 1""", stdout.getvalue(), re.X
-    )
-
-
-def test_rows_only_base_returns_a_dataframe_with_rows_only_in_base(spark, comparison1):
-    # require schema if contains only 1 row and contain field value as None
-    schema = StructType(
-        [
-            StructField("acct", LongType(), True),
-            StructField("date_fld", DateType(), True),
-            StructField("dollar_amt", LongType(), True),
-            StructField("float_fld", DoubleType(), True),
-            StructField("name", StringType(), True),
+            {"a": Decimal("1.3"), "b": Decimal("1.2"), "expected": False},
+            {"a": np.nan, "b": np.nan, "expected": True},
+            {"a": np.nan, "b": Decimal("1"), "expected": False},
+            {"a": Decimal("1"), "b": np.nan, "expected": False},
         ]
     )
-    expected_df = spark.createDataFrame(
-        [
-            Row(
-                acct=10000001239,
-                date_fld=datetime.date(2017, 1, 1),
-                dollar_amt=1,
-                float_fld=None,
-                name="Lucille Bluth",
-            )
-        ],
-        schema,
+    actual_out = columns_equal(df.a, df.b, abs_tol=0.001)
+    expect_out = df["expected"]
+    assert_series_equal(
+        expect_out.to_pandas(), actual_out.to_pandas(), check_names=False
     )
-    assert comparison1.rows_only_base.count() == 1
+
+
+@pandas_version
+def test_infinity_and_beyond():
+    # https://spark.apache.org/docs/latest/sql-ref-datatypes.html#positivenegative-infinity-semantics
+    # Positive/negative infinity multiplied by 0 returns NaN.
+    # Positive infinity sorts lower than NaN and higher than any other values.
+    # Negative infinity sorts lower than any other values.
+    df = ps.DataFrame(
+        [
+            {"a": np.inf, "b": np.inf, "expected": True},
+            {"a": -np.inf, "b": -np.inf, "expected": True},
+            {"a": -np.inf, "b": np.inf, "expected": True},
+            {"a": np.inf, "b": -np.inf, "expected": True},
+            {"a": 1, "b": 1, "expected": True},
+            {"a": 1, "b": 0, "expected": False},
+        ]
+    )
+    actual_out = columns_equal(df.a, df.b)
+    expect_out = df["expected"]
+    assert_series_equal(
+        expect_out.to_pandas(), actual_out.to_pandas(), check_names=False
+    )
+
+
+@pandas_version
+def test_compare_df_setter_bad():
+    df = ps.DataFrame([{"a": 1, "c": 2}, {"a": 2, "c": 2}])
+    with raises(TypeError, match="df1 must be a pyspark.pandas.frame.DataFrame"):
+        compare = SparkCompare("a", "a", ["a"])
+    with raises(ValueError, match="df1 must have all columns from join_columns"):
+        compare = SparkCompare(df, df.copy(), ["b"])
+    df_dupe = ps.DataFrame([{"a": 1, "b": 2}, {"a": 1, "b": 3}])
     assert (
-        expected_df.union(
-            comparison1.rows_only_base.select(
-                "acct", "date_fld", "dollar_amt", "float_fld", "name"
-            )
-        )
-        .distinct()
-        .count()
-        == 1
+        SparkCompare(df_dupe, df_dupe.copy(), ["a", "b"])
+        .df1.equals(df_dupe)
+        .all()
+        .all()
     )
 
 
-def test_rows_only_compare_returns_a_dataframe_with_rows_only_in_compare(
-    spark, comparison1
-):
-    expected_df = spark.createDataFrame(
+@pandas_version
+def test_compare_df_setter_good():
+    df1 = ps.DataFrame([{"a": 1, "b": 2}, {"a": 2, "b": 2}])
+    df2 = ps.DataFrame([{"A": 1, "B": 2}, {"A": 2, "B": 3}])
+    compare = SparkCompare(df1, df2, ["a"])
+    assert compare.df1.equals(df1).all().all()
+    assert compare.df2.equals(df2).all().all()
+    assert compare.join_columns == ["a"]
+    compare = SparkCompare(df1, df2, ["A", "b"])
+    assert compare.df1.equals(df1).all().all()
+    assert compare.df2.equals(df2).all().all()
+    assert compare.join_columns == ["a", "b"]
+
+
+@pandas_version
+def test_compare_df_setter_different_cases():
+    df1 = ps.DataFrame([{"a": 1, "b": 2}, {"a": 2, "b": 2}])
+    df2 = ps.DataFrame([{"A": 1, "b": 2}, {"A": 2, "b": 3}])
+    compare = SparkCompare(df1, df2, ["a"])
+    assert compare.df1.equals(df1).all().all()
+    assert compare.df2.equals(df2).all().all()
+
+
+@pandas_version
+def test_columns_overlap():
+    df1 = ps.DataFrame([{"a": 1, "b": 2}, {"a": 2, "b": 2}])
+    df2 = ps.DataFrame([{"a": 1, "b": 2}, {"a": 2, "b": 3}])
+    compare = SparkCompare(df1, df2, ["a"])
+    assert compare.df1_unq_columns() == set()
+    assert compare.df2_unq_columns() == set()
+    assert compare.intersect_columns() == {"a", "b"}
+
+
+@pandas_version
+def test_columns_no_overlap():
+    df1 = ps.DataFrame([{"a": 1, "b": 2, "c": "hi"}, {"a": 2, "b": 2, "c": "yo"}])
+    df2 = ps.DataFrame([{"a": 1, "b": 2, "d": "oh"}, {"a": 2, "b": 3, "d": "ya"}])
+    compare = SparkCompare(df1, df2, ["a"])
+    assert compare.df1_unq_columns() == {"c"}
+    assert compare.df2_unq_columns() == {"d"}
+    assert compare.intersect_columns() == {"a", "b"}
+
+
+@pandas_version
+def test_columns_maintain_order_through_set_operations():
+    df1 = ps.DataFrame(
         [
-            Row(
-                acct=10000001238,
-                dollar_amt=1.05,
-                name="Loose Seal Bluth",
-                float_fld=111.0,
-                accnt_purge=True,
-            )
+            (("A"), (0), (1), (2), (3), (4), (-2)),
+            (("B"), (0), (2), (2), (3), (4), (-3)),
+        ],
+        columns=["join", "f", "g", "b", "h", "a", "c"],
+    )
+    df2 = ps.DataFrame(
+        [
+            (("A"), (0), (1), (2), (-1), (4), (-3)),
+            (("B"), (1), (2), (3), (-1), (4), (-2)),
+        ],
+        columns=["join", "e", "h", "b", "a", "g", "d"],
+    )
+    compare = SparkCompare(df1, df2, ["join"])
+    assert list(compare.df1_unq_columns()) == ["f", "c"]
+    assert list(compare.df2_unq_columns()) == ["e", "d"]
+    assert list(compare.intersect_columns()) == ["join", "g", "b", "h", "a"]
+
+
+@pandas_version
+def test_10k_rows():
+    df1 = ps.DataFrame(np.random.randint(0, 100, size=(10000, 2)), columns=["b", "c"])
+    df1.reset_index(inplace=True)
+    df1.columns = ["a", "b", "c"]
+    df2 = df1.copy()
+    df2["b"] = df2["b"] + 0.1
+    compare_tol = SparkCompare(df1, df2, ["a"], abs_tol=0.2)
+    assert compare_tol.matches()
+    assert len(compare_tol.df1_unq_rows) == 0
+    assert len(compare_tol.df2_unq_rows) == 0
+    assert compare_tol.intersect_columns() == {"a", "b", "c"}
+    assert compare_tol.all_columns_match()
+    assert compare_tol.all_rows_overlap()
+    assert compare_tol.intersect_rows_match()
+
+    compare_no_tol = SparkCompare(df1, df2, ["a"])
+    assert not compare_no_tol.matches()
+    assert len(compare_no_tol.df1_unq_rows) == 0
+    assert len(compare_no_tol.df2_unq_rows) == 0
+    assert compare_no_tol.intersect_columns() == {"a", "b", "c"}
+    assert compare_no_tol.all_columns_match()
+    assert compare_no_tol.all_rows_overlap()
+    assert not compare_no_tol.intersect_rows_match()
+
+
+@pandas_version
+def test_subset(caplog):
+    caplog.set_level(logging.DEBUG)
+    df1 = ps.DataFrame([{"a": 1, "b": 2, "c": "hi"}, {"a": 2, "b": 2, "c": "yo"}])
+    df2 = ps.DataFrame([{"a": 1, "c": "hi"}])
+    comp = SparkCompare(df1, df2, ["a"])
+    assert comp.subset()
+    assert "Checking equality" in caplog.text
+
+
+@pandas_version
+def test_not_subset(caplog):
+    caplog.set_level(logging.INFO)
+    df1 = ps.DataFrame([{"a": 1, "b": 2, "c": "hi"}, {"a": 2, "b": 2, "c": "yo"}])
+    df2 = ps.DataFrame([{"a": 1, "b": 2, "c": "hi"}, {"a": 2, "b": 2, "c": "great"}])
+    comp = SparkCompare(df1, df2, ["a"])
+    assert not comp.subset()
+    assert "c: 1 / 2 (50.00%) match" in caplog.text
+
+
+@pandas_version
+def test_large_subset():
+    df1 = ps.DataFrame(np.random.randint(0, 100, size=(10000, 2)), columns=["b", "c"])
+    df1.reset_index(inplace=True)
+    df1.columns = ["a", "b", "c"]
+    df2 = df1[["a", "b"]].head(50).copy()
+    comp = SparkCompare(df1, df2, ["a"])
+    assert not comp.matches()
+    assert comp.subset()
+
+
+@pandas_version
+def test_string_joiner():
+    df1 = ps.DataFrame([{"ab": 1, "bc": 2}, {"ab": 2, "bc": 2}])
+    df2 = ps.DataFrame([{"ab": 1, "bc": 2}, {"ab": 2, "bc": 2}])
+    compare = SparkCompare(df1, df2, "ab")
+    assert compare.matches()
+
+
+@pandas_version
+def test_decimal_with_joins():
+    df1 = ps.DataFrame([{"a": Decimal("1"), "b": 2}, {"a": Decimal("2"), "b": 2}])
+    df2 = ps.DataFrame([{"a": 1, "b": 2}, {"a": 2, "b": 2}])
+    compare = SparkCompare(df1, df2, "a")
+    assert compare.matches()
+    assert compare.all_columns_match()
+    assert compare.all_rows_overlap()
+    assert compare.intersect_rows_match()
+
+
+@pandas_version
+def test_decimal_with_nulls():
+    df1 = ps.DataFrame([{"a": 1, "b": Decimal("2")}, {"a": 2, "b": Decimal("2")}])
+    df2 = ps.DataFrame([{"a": 1, "b": 2}, {"a": 2, "b": 2}, {"a": 3, "b": 2}])
+    compare = SparkCompare(df1, df2, "a")
+    assert not compare.matches()
+    assert compare.all_columns_match()
+    assert not compare.all_rows_overlap()
+    assert compare.intersect_rows_match()
+
+
+@pandas_version
+def test_strings_with_joins():
+    df1 = ps.DataFrame([{"a": "hi", "b": 2}, {"a": "bye", "b": 2}])
+    df2 = ps.DataFrame([{"a": "hi", "b": 2}, {"a": "bye", "b": 2}])
+    compare = SparkCompare(df1, df2, "a")
+    assert compare.matches()
+    assert compare.all_columns_match()
+    assert compare.all_rows_overlap()
+    assert compare.intersect_rows_match()
+
+
+@pandas_version
+def test_temp_column_name():
+    df1 = ps.DataFrame([{"a": "hi", "b": 2}, {"a": "bye", "b": 2}])
+    df2 = ps.DataFrame(
+        [{"a": "hi", "b": 2}, {"a": "bye", "b": 2}, {"a": "back fo mo", "b": 3}]
+    )
+    actual = temp_column_name(df1, df2)
+    assert actual == "_temp_0"
+
+
+@pandas_version
+def test_temp_column_name_one_has():
+    df1 = ps.DataFrame([{"_temp_0": "hi", "b": 2}, {"_temp_0": "bye", "b": 2}])
+    df2 = ps.DataFrame(
+        [{"a": "hi", "b": 2}, {"a": "bye", "b": 2}, {"a": "back fo mo", "b": 3}]
+    )
+    actual = temp_column_name(df1, df2)
+    assert actual == "_temp_1"
+
+
+@pandas_version
+def test_temp_column_name_both_have():
+    df1 = ps.DataFrame([{"_temp_0": "hi", "b": 2}, {"_temp_0": "bye", "b": 2}])
+    df2 = ps.DataFrame(
+        [
+            {"_temp_0": "hi", "b": 2},
+            {"_temp_0": "bye", "b": 2},
+            {"a": "back fo mo", "b": 3},
         ]
     )
+    actual = temp_column_name(df1, df2)
+    assert actual == "_temp_1"
 
-    assert comparison1.rows_only_compare.count() == 1
-    assert expected_df.union(comparison1.rows_only_compare).distinct().count() == 1
 
-
-def test_rows_both_mismatch_returns_a_dataframe_with_rows_where_variables_mismatched(
-    spark, comparison1
-):
-    expected_df = spark.createDataFrame(
+@pandas_version
+def test_temp_column_name_both_have():
+    df1 = ps.DataFrame([{"_temp_0": "hi", "b": 2}, {"_temp_0": "bye", "b": 2}])
+    df2 = ps.DataFrame(
         [
-            Row(
-                accnt_purge=False,
-                acct=10000001234,
-                date_fld=datetime.date(2017, 1, 1),
-                dollar_amt_base=123,
-                dollar_amt_compare=123.4,
-                dollar_amt_match=False,
-                float_fld_base=14530.1555,
-                float_fld_compare=14530.155,
-                float_fld_match=False,
-                name_base="George Maharis",
-                name_compare="George Michael Bluth",
-                name_match=False,
-            ),
-            Row(
-                accnt_purge=False,
-                acct=10000001235,
-                date_fld=datetime.date(2017, 1, 1),
-                dollar_amt_base=0,
-                dollar_amt_compare=0.45,
-                dollar_amt_match=False,
-                float_fld_base=1.0,
-                float_fld_compare=None,
-                float_fld_match=False,
-                name_base="Michael Bluth",
-                name_compare="Michael Bluth",
-                name_match=True,
-            ),
-            Row(
-                accnt_purge=False,
-                acct=10000001236,
-                date_fld=datetime.date(2017, 1, 1),
-                dollar_amt_base=1345,
-                dollar_amt_compare=1345.0,
-                dollar_amt_match=True,
-                float_fld_base=None,
-                float_fld_compare=1.0,
-                float_fld_match=False,
-                name_base="George Bluth",
-                name_compare="George Bluth",
-                name_match=True,
-            ),
+            {"_temp_0": "hi", "b": 2},
+            {"_temp_1": "bye", "b": 2},
+            {"a": "back fo mo", "b": 3},
         ]
     )
+    actual = temp_column_name(df1, df2)
+    assert actual == "_temp_2"
 
-    assert comparison1.rows_both_mismatch.count() == 3
-    assert expected_df.union(comparison1.rows_both_mismatch).distinct().count() == 3
 
-
-def test_rows_both_mismatch_only_includes_rows_with_true_mismatches_when_known_diffs_are_present(
-    spark, comparison_kd1
-):
-    expected_df = spark.createDataFrame(
+@pandas_version
+def test_temp_column_name_one_already():
+    df1 = ps.DataFrame([{"_temp_1": "hi", "b": 2}, {"_temp_1": "bye", "b": 2}])
+    df2 = ps.DataFrame(
         [
-            Row(
-                acct=10000001237,
-                acct_seq=0,
-                cd_base="0004",
-                cd_compare=4.0,
-                cd_match=True,
-                cd_match_type="KNOWN_DIFFERENCE",
-                open_dt_base=datetime.date(2017, 5, 4),
-                open_dt_compare=2017124,
-                open_dt_match=True,
-                open_dt_match_type="KNOWN_DIFFERENCE",
-                stat_cd_base="*2",
-                stat_cd_compare="V3",
-                stat_cd_match=False,
-                stat_cd_match_type="MISMATCH",
-            )
+            {"_temp_1": "hi", "b": 2},
+            {"_temp_1": "bye", "b": 2},
+            {"a": "back fo mo", "b": 3},
         ]
     )
-    assert comparison_kd1.rows_both_mismatch.count() == 1
-    assert expected_df.union(comparison_kd1.rows_both_mismatch).distinct().count() == 1
+    actual = temp_column_name(df1, df2)
+    assert actual == "_temp_0"
 
 
-def test_rows_both_all_returns_a_dataframe_with_all_rows_in_both_dataframes(
-    spark, comparison1
-):
-    expected_df = spark.createDataFrame(
-        [
-            Row(
-                accnt_purge=False,
-                acct=10000001234,
-                date_fld=datetime.date(2017, 1, 1),
-                dollar_amt_base=123,
-                dollar_amt_compare=123.4,
-                dollar_amt_match=False,
-                float_fld_base=14530.1555,
-                float_fld_compare=14530.155,
-                float_fld_match=False,
-                name_base="George Maharis",
-                name_compare="George Michael Bluth",
-                name_match=False,
-            ),
-            Row(
-                accnt_purge=False,
-                acct=10000001235,
-                date_fld=datetime.date(2017, 1, 1),
-                dollar_amt_base=0,
-                dollar_amt_compare=0.45,
-                dollar_amt_match=False,
-                float_fld_base=1.0,
-                float_fld_compare=None,
-                float_fld_match=False,
-                name_base="Michael Bluth",
-                name_compare="Michael Bluth",
-                name_match=True,
-            ),
-            Row(
-                accnt_purge=False,
-                acct=10000001236,
-                date_fld=datetime.date(2017, 1, 1),
-                dollar_amt_base=1345,
-                dollar_amt_compare=1345.0,
-                dollar_amt_match=True,
-                float_fld_base=None,
-                float_fld_compare=1.0,
-                float_fld_match=False,
-                name_base="George Bluth",
-                name_compare="George Bluth",
-                name_match=True,
-            ),
-            Row(
-                accnt_purge=False,
-                acct=10000001237,
-                date_fld=datetime.date(2017, 1, 1),
-                dollar_amt_base=123456,
-                dollar_amt_compare=123456.0,
-                dollar_amt_match=True,
-                float_fld_base=345.12,
-                float_fld_compare=345.12,
-                float_fld_match=True,
-                name_base="Bob Loblaw",
-                name_compare="Bob Loblaw",
-                name_match=True,
-            ),
-        ]
+### Duplicate testing!
+@pandas_version
+def test_simple_dupes_one_field():
+    df1 = ps.DataFrame([{"a": 1, "b": 2}, {"a": 1, "b": 2}])
+    df2 = ps.DataFrame([{"a": 1, "b": 2}, {"a": 1, "b": 2}])
+    compare = SparkCompare(df1, df2, join_columns=["a"])
+    assert compare.matches()
+    # Just render the report to make sure it renders.
+    t = compare.report()
+
+
+@pandas_version
+def test_simple_dupes_two_fields():
+    df1 = ps.DataFrame([{"a": 1, "b": 2}, {"a": 1, "b": 2, "c": 2}])
+    df2 = ps.DataFrame([{"a": 1, "b": 2}, {"a": 1, "b": 2, "c": 2}])
+    compare = SparkCompare(df1, df2, join_columns=["a", "b"])
+    assert compare.matches()
+    # Just render the report to make sure it renders.
+    t = compare.report()
+
+
+@pandas_version
+def test_simple_dupes_one_field_two_vals_1():
+    df1 = ps.DataFrame([{"a": 1, "b": 2}, {"a": 1, "b": 0}])
+    df2 = ps.DataFrame([{"a": 1, "b": 2}, {"a": 1, "b": 0}])
+    compare = SparkCompare(df1, df2, join_columns=["a"])
+    assert compare.matches()
+    # Just render the report to make sure it renders.
+    t = compare.report()
+
+
+@pandas_version
+def test_simple_dupes_one_field_two_vals_2():
+    df1 = ps.DataFrame([{"a": 1, "b": 2}, {"a": 1, "b": 0}])
+    df2 = ps.DataFrame([{"a": 1, "b": 2}, {"a": 2, "b": 0}])
+    compare = SparkCompare(df1, df2, join_columns=["a"])
+    assert not compare.matches()
+    assert len(compare.df1_unq_rows) == 1
+    assert len(compare.df2_unq_rows) == 1
+    assert len(compare.intersect_rows) == 1
+    # Just render the report to make sure it renders.
+    t = compare.report()
+
+
+@pandas_version
+def test_simple_dupes_one_field_three_to_two_vals():
+    df1 = ps.DataFrame([{"a": 1, "b": 2}, {"a": 1, "b": 0}, {"a": 1, "b": 0}])
+    df2 = ps.DataFrame([{"a": 1, "b": 2}, {"a": 1, "b": 0}])
+    compare = SparkCompare(df1, df2, join_columns=["a"])
+    assert not compare.matches()
+    assert len(compare.df1_unq_rows) == 1
+    assert len(compare.df2_unq_rows) == 0
+    assert len(compare.intersect_rows) == 2
+    # Just render the report to make sure it renders.
+    t = compare.report()
+
+    assert "(First 1 Columns)" in compare.report(column_count=1)
+    assert "(First 2 Columns)" in compare.report(column_count=2)
+
+
+@pandas_version
+def test_dupes_from_real_data():
+    data = """acct_id,acct_sfx_num,trxn_post_dt,trxn_post_seq_num,trxn_amt,trxn_dt,debit_cr_cd,cash_adv_trxn_comn_cntry_cd,mrch_catg_cd,mrch_pstl_cd,visa_mail_phn_cd,visa_rqstd_pmt_svc_cd,mc_pmt_facilitator_idn_num
+100,0,2017-06-17,1537019,30.64,2017-06-15,D,CAN,5812,M2N5P5,,,0.0
+200,0,2017-06-24,1022477,485.32,2017-06-22,D,USA,4511,7114,7.0,1,
+100,0,2017-06-17,1537039,2.73,2017-06-16,D,CAN,5812,M4J 1M9,,,0.0
+200,0,2017-06-29,1049223,22.41,2017-06-28,D,USA,4789,21211,,A,
+100,0,2017-06-17,1537029,34.05,2017-06-16,D,CAN,5812,M4E 2C7,,,0.0
+200,0,2017-06-29,1049213,9.12,2017-06-28,D,CAN,5814,0,,,
+100,0,2017-06-19,1646426,165.21,2017-06-17,D,CAN,5411,M4M 3H9,,,0.0
+200,0,2017-06-30,1233082,28.54,2017-06-29,D,USA,4121,94105,7.0,G,
+100,0,2017-06-19,1646436,17.87,2017-06-18,D,CAN,5812,M4J 1M9,,,0.0
+200,0,2017-06-30,1233092,24.39,2017-06-29,D,USA,4121,94105,7.0,G,
+100,0,2017-06-19,1646446,5.27,2017-06-17,D,CAN,5200,M4M 3G6,,,0.0
+200,0,2017-06-30,1233102,61.8,2017-06-30,D,CAN,4121,0,,,
+100,0,2017-06-20,1607573,41.99,2017-06-19,D,CAN,5661,M4C1M9,,,0.0
+200,0,2017-07-01,1009403,2.31,2017-06-29,D,USA,5814,22102,,F,
+100,0,2017-06-20,1607553,86.88,2017-06-19,D,CAN,4812,H2R3A8,,,0.0
+200,0,2017-07-01,1009423,5.5,2017-06-29,D,USA,5812,2903,,F,
+100,0,2017-06-20,1607563,25.17,2017-06-19,D,CAN,5641,M4C 1M9,,,0.0
+200,0,2017-07-01,1009433,214.12,2017-06-29,D,USA,3640,20170,,A,
+100,0,2017-06-20,1607593,1.67,2017-06-19,D,CAN,5814,M2N 6L7,,,0.0
+200,0,2017-07-01,1009393,2.01,2017-06-29,D,USA,5814,22102,,F,"""
+    df1 = ps.from_pandas(pd.read_csv(StringIO(data), sep=","))
+    df2 = df1.copy()
+    compare_acct = SparkCompare(df1, df2, join_columns=["acct_id"])
+    assert compare_acct.matches()
+    compare_unq = SparkCompare(
+        df1,
+        df2,
+        join_columns=["acct_id", "acct_sfx_num", "trxn_post_dt", "trxn_post_seq_num"],
     )
-
-    assert comparison1.rows_both_all.count() == 4
-    assert expected_df.union(comparison1.rows_both_all).distinct().count() == 4
-
-
-def test_rows_both_all_shows_known_diffs_flag_and_known_diffs_count_as_matches(
-    spark, comparison_kd1
-):
-    expected_df = spark.createDataFrame(
-        [
-            Row(
-                acct=10000001234,
-                acct_seq=0,
-                cd_base="0001",
-                cd_compare=1.0,
-                cd_match=True,
-                cd_match_type="KNOWN_DIFFERENCE",
-                open_dt_base=datetime.date(2017, 5, 1),
-                open_dt_compare=2017121,
-                open_dt_match=True,
-                open_dt_match_type="KNOWN_DIFFERENCE",
-                stat_cd_base="*2",
-                stat_cd_compare=None,
-                stat_cd_match=True,
-                stat_cd_match_type="KNOWN_DIFFERENCE",
-            ),
-            Row(
-                acct=10000001235,
-                acct_seq=0,
-                cd_base="0002",
-                cd_compare=2.0,
-                cd_match=True,
-                cd_match_type="KNOWN_DIFFERENCE",
-                open_dt_base=datetime.date(2017, 5, 2),
-                open_dt_compare=2017122,
-                open_dt_match=True,
-                open_dt_match_type="KNOWN_DIFFERENCE",
-                stat_cd_base="V1",
-                stat_cd_compare="V1",
-                stat_cd_match=True,
-                stat_cd_match_type="MATCH",
-            ),
-            Row(
-                acct=10000001236,
-                acct_seq=0,
-                cd_base="0003",
-                cd_compare=3.0,
-                cd_match=True,
-                cd_match_type="KNOWN_DIFFERENCE",
-                open_dt_base=datetime.date(2017, 5, 3),
-                open_dt_compare=2017123,
-                open_dt_match=True,
-                open_dt_match_type="KNOWN_DIFFERENCE",
-                stat_cd_base="V2",
-                stat_cd_compare="V2",
-                stat_cd_match=True,
-                stat_cd_match_type="MATCH",
-            ),
-            Row(
-                acct=10000001237,
-                acct_seq=0,
-                cd_base="0004",
-                cd_compare=4.0,
-                cd_match=True,
-                cd_match_type="KNOWN_DIFFERENCE",
-                open_dt_base=datetime.date(2017, 5, 4),
-                open_dt_compare=2017124,
-                open_dt_match=True,
-                open_dt_match_type="KNOWN_DIFFERENCE",
-                stat_cd_base="*2",
-                stat_cd_compare="V3",
-                stat_cd_match=False,
-                stat_cd_match_type="MISMATCH",
-            ),
-            Row(
-                acct=10000001238,
-                acct_seq=0,
-                cd_base="0005",
-                cd_compare=5.0,
-                cd_match=True,
-                cd_match_type="KNOWN_DIFFERENCE",
-                open_dt_base=datetime.date(2017, 5, 5),
-                open_dt_compare=2017125,
-                open_dt_match=True,
-                open_dt_match_type="KNOWN_DIFFERENCE",
-                stat_cd_base="*2",
-                stat_cd_compare=None,
-                stat_cd_match=True,
-                stat_cd_match_type="KNOWN_DIFFERENCE",
-            ),
-        ]
-    )
-
-    assert comparison_kd1.rows_both_all.count() == 5
-    assert expected_df.union(comparison_kd1.rows_both_all).distinct().count() == 5
+    assert compare_unq.matches()
+    # Just render the report to make sure it renders.
+    t = compare_acct.report()
+    r = compare_unq.report()
 
 
-def test_rows_both_all_returns_a_dataframe_with_all_rows_in_identical_dataframes(
-    spark, comparison2
-):
-    expected_df = spark.createDataFrame(
-        [
-            Row(
-                acct=10000001234,
-                date_fld_base=datetime.date(2017, 1, 1),
-                date_fld_compare=datetime.date(2017, 1, 1),
-                date_fld_match=True,
-                dollar_amt_base=123,
-                dollar_amt_compare=123,
-                dollar_amt_match=True,
-                float_fld_base=14530.1555,
-                float_fld_compare=14530.1555,
-                float_fld_match=True,
-                name_base="George Maharis",
-                name_compare="George Maharis",
-                name_match=True,
-            ),
-            Row(
-                acct=10000001235,
-                date_fld_base=datetime.date(2017, 1, 1),
-                date_fld_compare=datetime.date(2017, 1, 1),
-                date_fld_match=True,
-                dollar_amt_base=0,
-                dollar_amt_compare=0,
-                dollar_amt_match=True,
-                float_fld_base=1.0,
-                float_fld_compare=1.0,
-                float_fld_match=True,
-                name_base="Michael Bluth",
-                name_compare="Michael Bluth",
-                name_match=True,
-            ),
-            Row(
-                acct=10000001236,
-                date_fld_base=datetime.date(2017, 1, 1),
-                date_fld_compare=datetime.date(2017, 1, 1),
-                date_fld_match=True,
-                dollar_amt_base=1345,
-                dollar_amt_compare=1345,
-                dollar_amt_match=True,
-                float_fld_base=None,
-                float_fld_compare=None,
-                float_fld_match=True,
-                name_base="George Bluth",
-                name_compare="George Bluth",
-                name_match=True,
-            ),
-            Row(
-                acct=10000001237,
-                date_fld_base=datetime.date(2017, 1, 1),
-                date_fld_compare=datetime.date(2017, 1, 1),
-                date_fld_match=True,
-                dollar_amt_base=123456,
-                dollar_amt_compare=123456,
-                dollar_amt_match=True,
-                float_fld_base=345.12,
-                float_fld_compare=345.12,
-                float_fld_match=True,
-                name_base="Bob Loblaw",
-                name_compare="Bob Loblaw",
-                name_match=True,
-            ),
-            Row(
-                acct=10000001239,
-                date_fld_base=datetime.date(2017, 1, 1),
-                date_fld_compare=datetime.date(2017, 1, 1),
-                date_fld_match=True,
-                dollar_amt_base=1,
-                dollar_amt_compare=1,
-                dollar_amt_match=True,
-                float_fld_base=None,
-                float_fld_compare=None,
-                float_fld_match=True,
-                name_base="Lucille Bluth",
-                name_compare="Lucille Bluth",
-                name_match=True,
-            ),
-        ]
-    )
+@pandas_version
+def test_strings_with_joins_with_ignore_spaces():
+    df1 = ps.DataFrame([{"a": "hi", "b": " A"}, {"a": "bye", "b": "A"}])
+    df2 = ps.DataFrame([{"a": "hi", "b": "A"}, {"a": "bye", "b": "A "}])
+    compare = SparkCompare(df1, df2, "a", ignore_spaces=False)
+    assert not compare.matches()
+    assert compare.all_columns_match()
+    assert compare.all_rows_overlap()
+    assert not compare.intersect_rows_match()
 
-    assert comparison2.rows_both_all.count() == 5
-    assert expected_df.union(comparison2.rows_both_all).distinct().count() == 5
+    compare = SparkCompare(df1, df2, "a", ignore_spaces=True)
+    assert compare.matches()
+    assert compare.all_columns_match()
+    assert compare.all_rows_overlap()
+    assert compare.intersect_rows_match()
 
 
-def test_rows_both_all_returns_all_rows_in_both_dataframes_for_differently_named_columns(
-    spark, comparison3
-):
-    expected_df = spark.createDataFrame(
-        [
-            Row(
-                accnt_purge=False,
-                acct=10000001234,
-                date_fld_base=datetime.date(2017, 1, 1),
-                date_fld_compare=datetime.date(2017, 1, 1),
-                date_fld_match=True,
-                dollar_amt_base=123,
-                dollar_amt_compare=123.4,
-                dollar_amt_match=False,
-                float_fld_base=14530.1555,
-                float_fld_compare=14530.155,
-                float_fld_match=False,
-                name_base="George Maharis",
-                name_compare="George Michael Bluth",
-                name_match=False,
-            ),
-            Row(
-                accnt_purge=False,
-                acct=10000001235,
-                date_fld_base=datetime.date(2017, 1, 1),
-                date_fld_compare=datetime.date(2017, 1, 1),
-                date_fld_match=True,
-                dollar_amt_base=0,
-                dollar_amt_compare=0.45,
-                dollar_amt_match=False,
-                float_fld_base=1.0,
-                float_fld_compare=1.0,
-                float_fld_match=True,
-                name_base="Michael Bluth",
-                name_compare="Michael Bluth",
-                name_match=True,
-            ),
-            Row(
-                accnt_purge=False,
-                acct=10000001236,
-                date_fld_base=datetime.date(2017, 1, 1),
-                date_fld_compare=datetime.date(2017, 1, 1),
-                date_fld_match=True,
-                dollar_amt_base=1345,
-                dollar_amt_compare=1345.0,
-                dollar_amt_match=True,
-                float_fld_base=None,
-                float_fld_compare=None,
-                float_fld_match=True,
-                name_base="George Bluth",
-                name_compare="George Bluth",
-                name_match=True,
-            ),
-            Row(
-                accnt_purge=False,
-                acct=10000001237,
-                date_fld_base=datetime.date(2017, 1, 1),
-                date_fld_compare=datetime.date(2017, 1, 1),
-                date_fld_match=True,
-                dollar_amt_base=123456,
-                dollar_amt_compare=123456.0,
-                dollar_amt_match=True,
-                float_fld_base=345.12,
-                float_fld_compare=345.12,
-                float_fld_match=True,
-                name_base="Bob Loblaw",
-                name_compare="Bob Loblaw",
-                name_match=True,
-            ),
-            Row(
-                accnt_purge=True,
-                acct=10000001239,
-                date_fld_base=datetime.date(2017, 1, 1),
-                date_fld_compare=datetime.date(2017, 1, 1),
-                date_fld_match=True,
-                dollar_amt_base=1,
-                dollar_amt_compare=1.05,
-                dollar_amt_match=False,
-                float_fld_base=None,
-                float_fld_compare=None,
-                float_fld_match=True,
-                name_base="Lucille Bluth",
-                name_compare="Lucille Bluth",
-                name_match=True,
-            ),
-        ]
-    )
+@pandas_version
+def test_strings_with_joins_with_ignore_case():
+    df1 = ps.DataFrame([{"a": "hi", "b": "a"}, {"a": "bye", "b": "A"}])
+    df2 = ps.DataFrame([{"a": "hi", "b": "A"}, {"a": "bye", "b": "a"}])
+    compare = SparkCompare(df1, df2, "a", ignore_case=False)
+    assert not compare.matches()
+    assert compare.all_columns_match()
+    assert compare.all_rows_overlap()
+    assert not compare.intersect_rows_match()
 
-    assert comparison3.rows_both_all.count() == 5
-    assert expected_df.union(comparison3.rows_both_all).distinct().count() == 5
+    compare = SparkCompare(df1, df2, "a", ignore_case=True)
+    assert compare.matches()
+    assert compare.all_columns_match()
+    assert compare.all_rows_overlap()
+    assert compare.intersect_rows_match()
 
 
-def test_columns_with_unequal_values_text_is_aligned(comparison4):
-    stdout = io.StringIO()
+@pandas_version
+def test_decimal_with_joins_with_ignore_spaces():
+    df1 = ps.DataFrame([{"a": 1, "b": " A"}, {"a": 2, "b": "A"}])
+    df2 = ps.DataFrame([{"a": 1, "b": "A"}, {"a": 2, "b": "A "}])
+    compare = SparkCompare(df1, df2, "a", ignore_spaces=False)
+    assert not compare.matches()
+    assert compare.all_columns_match()
+    assert compare.all_rows_overlap()
+    assert not compare.intersect_rows_match()
 
-    comparison4.report(file=stdout)
-    stdout.seek(0)  # Back up to the beginning of the stream
-
-    text_alignment_validator(
-        report=stdout,
-        section_start="****** Columns with Unequal Values ******",
-        section_end="\n",
-        left_indices=(1, 2, 3, 4),
-        right_indices=(5, 6),
-        column_regexes=[
-            r"""(Base\sColumn\sName) \s+ (Compare\sColumn\sName) \s+ (Base\sDtype) \s+ (Compare\sDtype) \s+
-                (\#\sMatches) \s+ (\#\sMismatches)""",
-            r"""(-+) \s+ (-+) \s+ (-+) \s+ (-+) \s+ (-+) \s+ (-+)""",
-            r"""(dollar_amt) \s+ (dollar_amt) \s+ (bigint) \s+ (double) \s+ (2) \s+ (2)""",
-            r"""(float_fld) \s+ (float_fld) \s+ (double) \s+ (double) \s+ (1) \s+ (3)""",
-            r"""(super_duper_big_long_name) \s+ (name) \s+ (string) \s+ (string) \s+ (3) \s+ (1)\s*""",
-        ],
-    )
+    compare = SparkCompare(df1, df2, "a", ignore_spaces=True)
+    assert compare.matches()
+    assert compare.all_columns_match()
+    assert compare.all_rows_overlap()
+    assert compare.intersect_rows_match()
 
 
-def test_columns_with_unequal_values_text_is_aligned_with_known_differences(
-    comparison_kd1,
-):
-    stdout = io.StringIO()
+@pandas_version
+def test_decimal_with_joins_with_ignore_case():
+    df1 = ps.DataFrame([{"a": 1, "b": "a"}, {"a": 2, "b": "A"}])
+    df2 = ps.DataFrame([{"a": 1, "b": "A"}, {"a": 2, "b": "a"}])
+    compare = SparkCompare(df1, df2, "a", ignore_case=False)
+    assert not compare.matches()
+    assert compare.all_columns_match()
+    assert compare.all_rows_overlap()
+    assert not compare.intersect_rows_match()
 
-    comparison_kd1.report(file=stdout)
-    stdout.seek(0)  # Back up to the beginning of the stream
-
-    text_alignment_validator(
-        report=stdout,
-        section_start="****** Columns with Unequal Values ******",
-        section_end="\n",
-        left_indices=(1, 2, 3, 4),
-        right_indices=(5, 6, 7),
-        column_regexes=[
-            r"""(Base\sColumn\sName) \s+ (Compare\sColumn\sName) \s+ (Base\sDtype) \s+ (Compare\sDtype) \s+
-                (\#\sMatches) \s+ (\#\sKnown\sDiffs) \s+ (\#\sMismatches)""",
-            r"""(-+) \s+ (-+) \s+ (-+) \s+ (-+) \s+ (-+) \s+ (-+) \s+ (-+)""",
-            r"""(stat_cd) \s+ (STATC) \s+ (string) \s+ (string) \s+ (2) \s+ (2) \s+ (1)""",
-            r"""(open_dt) \s+ (ACCOUNT_OPEN) \s+ (date) \s+ (bigint) \s+ (0) \s+ (5) \s+ (0)""",
-            r"""(cd) \s+ (CODE) \s+ (string) \s+ (double) \s+ (0) \s+ (5) \s+ (0)\s*""",
-        ],
-    )
+    compare = SparkCompare(df1, df2, "a", ignore_case=True)
+    assert compare.matches()
+    assert compare.all_columns_match()
+    assert compare.all_rows_overlap()
+    assert compare.intersect_rows_match()
 
 
-def test_columns_with_unequal_values_text_is_aligned_with_custom_known_differences(
-    comparison_kd2,
-):
-    stdout = io.StringIO()
+@pandas_version
+def test_joins_with_ignore_spaces():
+    df1 = ps.DataFrame([{"a": 1, "b": " A"}, {"a": 2, "b": "A"}])
+    df2 = ps.DataFrame([{"a": 1, "b": "A"}, {"a": 2, "b": "A "}])
 
-    comparison_kd2.report(file=stdout)
-    stdout.seek(0)  # Back up to the beginning of the stream
-
-    text_alignment_validator(
-        report=stdout,
-        section_start="****** Columns with Unequal Values ******",
-        section_end="\n",
-        left_indices=(1, 2, 3, 4),
-        right_indices=(5, 6, 7),
-        column_regexes=[
-            r"""(Base\sColumn\sName) \s+ (Compare\sColumn\sName) \s+ (Base\sDtype) \s+ (Compare\sDtype) \s+
-                (\#\sMatches) \s+ (\#\sKnown\sDiffs) \s+ (\#\sMismatches)""",
-            r"""(-+) \s+ (-+) \s+ (-+) \s+ (-+) \s+ (-+) \s+ (-+) \s+ (-+)""",
-            r"""(stat_cd) \s+ (STATC) \s+ (string) \s+ (string) \s+ (2) \s+ (2) \s+ (1)""",
-            r"""(open_dt) \s+ (ACCOUNT_OPEN) \s+ (date) \s+ (bigint) \s+ (0) \s+ (0) \s+ (5)""",
-            r"""(cd) \s+ (CODE) \s+ (string) \s+ (double) \s+ (0) \s+ (5) \s+ (0)\s*""",
-        ],
-    )
+    compare = SparkCompare(df1, df2, "a", ignore_spaces=True)
+    assert compare.matches()
+    assert compare.all_columns_match()
+    assert compare.all_rows_overlap()
+    assert compare.intersect_rows_match()
 
 
-def test_columns_with_unequal_values_text_is_aligned_for_decimals(comparison_decimal):
-    stdout = io.StringIO()
+@pandas_version
+def test_joins_with_ignore_case():
+    df1 = ps.DataFrame([{"a": 1, "b": "a"}, {"a": 2, "b": "A"}])
+    df2 = ps.DataFrame([{"a": 1, "b": "A"}, {"a": 2, "b": "a"}])
 
-    comparison_decimal.report(file=stdout)
-    stdout.seek(0)  # Back up to the beginning of the stream
-
-    text_alignment_validator(
-        report=stdout,
-        section_start="****** Columns with Unequal Values ******",
-        section_end="\n",
-        left_indices=(1, 2, 3, 4),
-        right_indices=(5, 6),
-        column_regexes=[
-            r"""(Base\sColumn\sName) \s+ (Compare\sColumn\sName) \s+ (Base\sDtype) \s+ (Compare\sDtype) \s+
-                (\#\sMatches) \s+ (\#\sMismatches)""",
-            r"""(-+) \s+ (-+) \s+ (-+) \s+ (-+) \s+ (-+) \s+ (-+)""",
-            r"""(dollar_amt) \s+ (dollar_amt) \s+ (decimal\(8,2\)) \s+ (double) \s+ (1) \s+ (1)""",
-        ],
-    )
+    compare = SparkCompare(df1, df2, "a", ignore_case=True)
+    assert compare.matches()
+    assert compare.all_columns_match()
+    assert compare.all_rows_overlap()
+    assert compare.intersect_rows_match()
 
 
-def test_schema_differences_text_is_aligned(comparison4):
-    stdout = io.StringIO()
+@pandas_version
+def test_strings_with_ignore_spaces_and_join_columns():
+    df1 = ps.DataFrame([{"a": "hi", "b": "A"}, {"a": "bye", "b": "A"}])
+    df2 = ps.DataFrame([{"a": " hi ", "b": "A"}, {"a": " bye ", "b": "A"}])
+    compare = SparkCompare(df1, df2, "a", ignore_spaces=False)
+    assert not compare.matches()
+    assert compare.all_columns_match()
+    assert not compare.all_rows_overlap()
+    assert compare.count_matching_rows() == 0
 
-    comparison4.report(file=stdout)
-    comparison4.report()
-    stdout.seek(0)  # Back up to the beginning of the stream
-
-    text_alignment_validator(
-        report=stdout,
-        section_start="****** Schema Differences ******",
-        section_end="\n",
-        left_indices=(1, 2, 3, 4),
-        right_indices=(),
-        column_regexes=[
-            r"""(Base\sColumn\sName) \s+ (Compare\sColumn\sName) \s+ (Base\sDtype) \s+ (Compare\sDtype)""",
-            r"""(-+) \s+ (-+) \s+ (-+) \s+ (-+)""",
-            r"""(dollar_amt) \s+ (dollar_amt) \s+ (bigint) \s+ (double)""",
-        ],
-    )
+    compare = SparkCompare(df1, df2, "a", ignore_spaces=True)
+    assert compare.matches()
+    assert compare.all_columns_match()
+    assert compare.all_rows_overlap()
+    assert compare.intersect_rows_match()
+    assert compare.count_matching_rows() == 2
 
 
-def test_schema_differences_text_is_aligned_for_decimals(comparison_decimal):
-    stdout = io.StringIO()
+@pandas_version
+def test_integers_with_ignore_spaces_and_join_columns():
+    df1 = ps.DataFrame([{"a": 1, "b": "A"}, {"a": 2, "b": "A"}])
+    df2 = ps.DataFrame([{"a": 1, "b": "A"}, {"a": 2, "b": "A"}])
+    compare = SparkCompare(df1, df2, "a", ignore_spaces=False)
+    assert compare.matches()
+    assert compare.all_columns_match()
+    assert compare.all_rows_overlap()
+    assert compare.intersect_rows_match()
+    assert compare.count_matching_rows() == 2
 
-    comparison_decimal.report(file=stdout)
-    stdout.seek(0)  # Back up to the beginning of the stream
-
-    text_alignment_validator(
-        report=stdout,
-        section_start="****** Schema Differences ******",
-        section_end="\n",
-        left_indices=(1, 2, 3, 4),
-        right_indices=(),
-        column_regexes=[
-            r"""(Base\sColumn\sName) \s+ (Compare\sColumn\sName) \s+ (Base\sDtype) \s+ (Compare\sDtype)""",
-            r"""(-+) \s+ (-+) \s+ (-+) \s+ (-+)""",
-            r"""(dollar_amt) \s+ (dollar_amt) \s+ (decimal\(8,2\)) \s+ (double)""",
-        ],
-    )
-
-
-def test_base_only_columns_text_is_aligned(comparison4):
-    stdout = io.StringIO()
-
-    comparison4.report(file=stdout)
-    stdout.seek(0)  # Back up to the beginning of the stream
-
-    text_alignment_validator(
-        report=stdout,
-        section_start="****** Columns In Base Only ******",
-        section_end="\n",
-        left_indices=(1, 2),
-        right_indices=(),
-        column_regexes=[
-            r"""(Column\sName) \s+ (Dtype)""",
-            r"""(-+) \s+ (-+)""",
-            r"""(date_fld) \s+ (date)""",
-        ],
-    )
+    compare = SparkCompare(df1, df2, "a", ignore_spaces=True)
+    assert compare.matches()
+    assert compare.all_columns_match()
+    assert compare.all_rows_overlap()
+    assert compare.intersect_rows_match()
+    assert compare.count_matching_rows() == 2
 
 
-def test_compare_only_columns_text_is_aligned(comparison4):
-    stdout = io.StringIO()
-
-    comparison4.report(file=stdout)
-    stdout.seek(0)  # Back up to the beginning of the stream
-
-    text_alignment_validator(
-        report=stdout,
-        section_start="****** Columns In Compare Only ******",
-        section_end="\n",
-        left_indices=(1, 2),
-        right_indices=(),
-        column_regexes=[
-            r"""(Column\sName) \s+ (Dtype)""",
-            r"""(-+) \s+ (-+)""",
-            r"""(accnt_purge) \s+ (boolean)""",
-        ],
-    )
-
-
-def text_alignment_validator(
-    report, section_start, section_end, left_indices, right_indices, column_regexes
-):
-    r"""Check to make sure that report output columns are vertically aligned.
-
-    Parameters
-    ----------
-    report: An iterable returning lines of report output to be validated.
-    section_start: A string that represents the beginning of the section to be validated.
-    section_end: A string that represents the end of the section to be validated.
-    left_indices: The match group indexes (starting with 1) that should be left-aligned
-        in the output column.
-    right_indices: The match group indexes (starting with 1) that should be right-aligned
-        in the output column.
-    column_regexes: A list of regular expressions representing the expected output, with
-        each column enclosed with parentheses to return a match. The regular expression will
-        use the "X" flag, so it may contain whitespace, and any whitespace to be matched
-        should be explicitly given with \s. The first line will represent the alignments
-        that are expected in the following lines. The number of match groups should cover
-        all of the indices given in left/right_indices.
-
-    Runs assertions for every match group specified by left/right_indices to ensure that
-    all lines past the first are either left- or right-aligned with the same match group
-    on the first line.
+@pandas_version
+def test_sample_mismatch():
+    data1 = """acct_id,dollar_amt,name,float_fld,date_fld
+    10000001234,123.45,George Maharis,14530.1555,2017-01-01
+    10000001235,0.45,Michael Bluth,1,2017-01-01
+    10000001236,1345,George Bluth,,2017-01-01
+    10000001237,123456,Bob Loblaw,345.12,2017-01-01
+    10000001239,1.05,Lucille Bluth,,2017-01-01
+    10000001240,123.45,George Maharis,14530.1555,2017-01-02
     """
 
-    at_column_section = False
-    processed_first_line = False
-    match_positions = [None] * (len(left_indices + right_indices) + 1)
+    data2 = """acct_id,dollar_amt,name,float_fld,date_fld
+    10000001234,123.4,George Michael Bluth,14530.155,
+    10000001235,0.45,Michael Bluth,,
+    10000001236,1345,George Bluth,1,
+    10000001237,123456,Robert Loblaw,345.12,
+    10000001238,1.05,Loose Seal Bluth,111,
+    10000001240,123.45,George Maharis,14530.1555,2017-01-02
+    """
 
-    for line in report:
-        if at_column_section:
-            if line == section_end:  # Detect end of section and stop
-                break
+    df1 = ps.from_pandas(pd.read_csv(StringIO(data1), sep=","))
+    df2 = ps.from_pandas(pd.read_csv(StringIO(data2), sep=","))
 
-            if (
-                not processed_first_line
-            ):  # First line in section - capture text start/end positions
-                matches = re.search(column_regexes[0], line, re.X)
-                assert matches is not None  # Make sure we found at least this...
+    compare = SparkCompare(df1, df2, "acct_id")
 
-                for n in left_indices:
-                    match_positions[n] = matches.start(n)
-                for n in right_indices:
-                    match_positions[n] = matches.end(n)
-                processed_first_line = True
-            else:  # Match the stuff after the header text
-                match = None
-                for regex in column_regexes[1:]:
-                    match = re.search(regex, line, re.X)
-                    if match:
-                        break
+    output = compare.sample_mismatch(column="name", sample_count=1)
+    assert output.shape[0] == 1
+    assert (output.name_df1 != output.name_df2).all()
 
-                if not match:
-                    raise AssertionError(f'Did not find a match for line: "{line}"')
+    output = compare.sample_mismatch(column="name", sample_count=2)
+    assert output.shape[0] == 2
+    assert (output.name_df1 != output.name_df2).all()
 
-                for n in left_indices:
-                    assert match_positions[n] == match.start(n)
-                for n in right_indices:
-                    assert match_positions[n] == match.end(n)
-
-        if not at_column_section and section_start in line:
-            at_column_section = True
+    output = compare.sample_mismatch(column="name", sample_count=3)
+    assert output.shape[0] == 2
+    assert (output.name_df1 != output.name_df2).all()
 
 
-def test_unicode_columns(spark_session):
-    df1 = spark_session.createDataFrame(
-        [
-            (1, "foo", "test"),
-            (2, "bar", "test"),
-        ],
-        ["id", "例", "予測対象日"],
+@pandas_version
+def test_all_mismatch_not_ignore_matching_cols_no_cols_matching():
+    data1 = """acct_id,dollar_amt,name,float_fld,date_fld
+    10000001234,123.45,George Maharis,14530.1555,2017-01-01
+    10000001235,0.45,Michael Bluth,1,2017-01-01
+    10000001236,1345,George Bluth,,2017-01-01
+    10000001237,123456,Bob Loblaw,345.12,2017-01-01
+    10000001239,1.05,Lucille Bluth,,2017-01-01
+    10000001240,123.45,George Maharis,14530.1555,2017-01-02
+    """
+
+    data2 = """acct_id,dollar_amt,name,float_fld,date_fld
+    10000001234,123.4,George Michael Bluth,14530.155,
+    10000001235,0.45,Michael Bluth,,
+    10000001236,1345,George Bluth,1,
+    10000001237,123456,Robert Loblaw,345.12,
+    10000001238,1.05,Loose Seal Bluth,111,
+    10000001240,123.45,George Maharis,14530.1555,2017-01-02
+    """
+    df1 = ps.from_pandas(pd.read_csv(StringIO(data1), sep=","))
+    df2 = ps.from_pandas(pd.read_csv(StringIO(data2), sep=","))
+    compare = SparkCompare(df1, df2, "acct_id")
+
+    output = compare.all_mismatch()
+    assert output.shape[0] == 4
+    assert output.shape[1] == 10
+
+    assert (output.name_df1 != output.name_df2).values.sum() == 2
+    assert (~(output.name_df1 != output.name_df2)).values.sum() == 2
+
+    assert (output.dollar_amt_df1 != output.dollar_amt_df2).values.sum() == 1
+    assert (~(output.dollar_amt_df1 != output.dollar_amt_df2)).values.sum() == 3
+
+    assert (output.float_fld_df1 != output.float_fld_df2).values.sum() == 3
+    assert (~(output.float_fld_df1 != output.float_fld_df2)).values.sum() == 1
+
+    assert (output.date_fld_df1 != output.date_fld_df2).values.sum() == 4
+    assert (~(output.date_fld_df1 != output.date_fld_df2)).values.sum() == 0
+
+
+@pandas_version
+def test_all_mismatch_not_ignore_matching_cols_some_cols_matching():
+    # Columns dollar_amt and name are matching
+    data1 = """acct_id,dollar_amt,name,float_fld,date_fld
+        10000001234,123.45,George Maharis,14530.1555,2017-01-01
+        10000001235,0.45,Michael Bluth,1,2017-01-01
+        10000001236,1345,George Bluth,,2017-01-01
+        10000001237,123456,Bob Loblaw,345.12,2017-01-01
+        10000001239,1.05,Lucille Bluth,,2017-01-01
+        10000001240,123.45,George Maharis,14530.1555,2017-01-02
+        """
+
+    data2 = """acct_id,dollar_amt,name,float_fld,date_fld
+        10000001234,123.45,George Maharis,14530.155,
+        10000001235,0.45,Michael Bluth,,
+        10000001236,1345,George Bluth,1,
+        10000001237,123456,Bob Loblaw,345.12,
+        10000001238,1.05,Lucille Bluth,111,
+        10000001240,123.45,George Maharis,14530.1555,2017-01-02
+        """
+    df1 = ps.from_pandas(pd.read_csv(StringIO(data1), sep=","))
+    df2 = ps.from_pandas(pd.read_csv(StringIO(data2), sep=","))
+    compare = SparkCompare(df1, df2, "acct_id")
+
+    output = compare.all_mismatch()
+    assert output.shape[0] == 4
+    assert output.shape[1] == 10
+
+    assert (output.name_df1 != output.name_df2).values.sum() == 0
+    assert (~(output.name_df1 != output.name_df2)).values.sum() == 4
+
+    assert (output.dollar_amt_df1 != output.dollar_amt_df2).values.sum() == 0
+    assert (~(output.dollar_amt_df1 != output.dollar_amt_df2)).values.sum() == 4
+
+    assert (output.float_fld_df1 != output.float_fld_df2).values.sum() == 3
+    assert (~(output.float_fld_df1 != output.float_fld_df2)).values.sum() == 1
+
+    assert (output.date_fld_df1 != output.date_fld_df2).values.sum() == 4
+    assert (~(output.date_fld_df1 != output.date_fld_df2)).values.sum() == 0
+
+
+@pandas_version
+def test_all_mismatch_ignore_matching_cols_some_cols_matching_diff_rows():
+    # Case where there are rows on either dataset which don't match up.
+    # Columns dollar_amt and name are matching
+    data1 = """acct_id,dollar_amt,name,float_fld,date_fld
+    10000001234,123.45,George Maharis,14530.1555,2017-01-01
+    10000001235,0.45,Michael Bluth,1,2017-01-01
+    10000001236,1345,George Bluth,,2017-01-01
+    10000001237,123456,Bob Loblaw,345.12,2017-01-01
+    10000001239,1.05,Lucille Bluth,,2017-01-01
+    10000001240,123.45,George Maharis,14530.1555,2017-01-02
+    10000001241,1111.05,Lucille Bluth,
+    """
+
+    data2 = """acct_id,dollar_amt,name,float_fld,date_fld
+    10000001234,123.45,George Maharis,14530.155,
+    10000001235,0.45,Michael Bluth,,
+    10000001236,1345,George Bluth,1,
+    10000001237,123456,Bob Loblaw,345.12,
+    10000001238,1.05,Lucille Bluth,111,
+    """
+    df1 = ps.from_pandas(pd.read_csv(StringIO(data1), sep=","))
+    df2 = ps.from_pandas(pd.read_csv(StringIO(data2), sep=","))
+    compare = SparkCompare(df1, df2, "acct_id")
+
+    output = compare.all_mismatch(ignore_matching_cols=True)
+
+    assert output.shape[0] == 4
+    assert output.shape[1] == 6
+
+    assert (output.float_fld_df1 != output.float_fld_df2).values.sum() == 3
+    assert (~(output.float_fld_df1 != output.float_fld_df2)).values.sum() == 1
+
+    assert (output.date_fld_df1 != output.date_fld_df2).values.sum() == 4
+    assert (~(output.date_fld_df1 != output.date_fld_df2)).values.sum() == 0
+
+    assert not ("name_df1" in output and "name_df2" in output)
+    assert not ("dollar_amt_df1" in output and "dollar_amt_df1" in output)
+
+
+@pandas_version
+def test_all_mismatch_ignore_matching_cols_some_calls_matching():
+    # Columns dollar_amt and name are matching
+    data1 = """acct_id,dollar_amt,name,float_fld,date_fld
+    10000001234,123.45,George Maharis,14530.1555,2017-01-01
+    10000001235,0.45,Michael Bluth,1,2017-01-01
+    10000001236,1345,George Bluth,,2017-01-01
+    10000001237,123456,Bob Loblaw,345.12,2017-01-01
+    10000001239,1.05,Lucille Bluth,,2017-01-01
+    10000001240,123.45,George Maharis,14530.1555,2017-01-02
+    """
+
+    data2 = """acct_id,dollar_amt,name,float_fld,date_fld
+    10000001234,123.45,George Maharis,14530.155,
+    10000001235,0.45,Michael Bluth,,
+    10000001236,1345,George Bluth,1,
+    10000001237,123456,Bob Loblaw,345.12,
+    10000001238,1.05,Lucille Bluth,111,
+    10000001240,123.45,George Maharis,14530.1555,2017-01-02
+    """
+    df1 = ps.from_pandas(pd.read_csv(StringIO(data1), sep=","))
+    df2 = ps.from_pandas(pd.read_csv(StringIO(data2), sep=","))
+    compare = SparkCompare(df1, df2, "acct_id")
+
+    output = compare.all_mismatch(ignore_matching_cols=True)
+
+    assert output.shape[0] == 4
+    assert output.shape[1] == 6
+
+    assert (output.float_fld_df1 != output.float_fld_df2).values.sum() == 3
+    assert (~(output.float_fld_df1 != output.float_fld_df2)).values.sum() == 1
+
+    assert (output.date_fld_df1 != output.date_fld_df2).values.sum() == 4
+    assert (~(output.date_fld_df1 != output.date_fld_df2)).values.sum() == 0
+
+    assert not ("name_df1" in output and "name_df2" in output)
+    assert not ("dollar_amt_df1" in output and "dollar_amt_df1" in output)
+
+
+@pandas_version
+def test_all_mismatch_ignore_matching_cols_no_cols_matching():
+    data1 = """acct_id,dollar_amt,name,float_fld,date_fld
+    10000001234,123.45,George Maharis,14530.1555,2017-01-01
+    10000001235,0.45,Michael Bluth,1,2017-01-01
+    10000001236,1345,George Bluth,,2017-01-01
+    10000001237,123456,Bob Loblaw,345.12,2017-01-01
+    10000001239,1.05,Lucille Bluth,,2017-01-01
+    10000001240,123.45,George Maharis,14530.1555,2017-01-02
+    """
+
+    data2 = """acct_id,dollar_amt,name,float_fld,date_fld
+    10000001234,123.4,George Michael Bluth,14530.155,
+    10000001235,0.45,Michael Bluth,,
+    10000001236,1345,George Bluth,1,
+    10000001237,123456,Robert Loblaw,345.12,
+    10000001238,1.05,Loose Seal Bluth,111,
+    10000001240,123.45,George Maharis,14530.1555,2017-01-02
+    """
+    df1 = ps.from_pandas(pd.read_csv(StringIO(data1), sep=","))
+    df2 = ps.from_pandas(pd.read_csv(StringIO(data2), sep=","))
+    compare = SparkCompare(df1, df2, "acct_id")
+
+    output = compare.all_mismatch()
+    assert output.shape[0] == 4
+    assert output.shape[1] == 10
+
+    assert (output.name_df1 != output.name_df2).values.sum() == 2
+    assert (~(output.name_df1 != output.name_df2)).values.sum() == 2
+
+    assert (output.dollar_amt_df1 != output.dollar_amt_df2).values.sum() == 1
+    assert (~(output.dollar_amt_df1 != output.dollar_amt_df2)).values.sum() == 3
+
+    assert (output.float_fld_df1 != output.float_fld_df2).values.sum() == 3
+    assert (~(output.float_fld_df1 != output.float_fld_df2)).values.sum() == 1
+
+    assert (output.date_fld_df1 != output.date_fld_df2).values.sum() == 4
+    assert (~(output.date_fld_df1 != output.date_fld_df2)).values.sum() == 0
+
+
+@pandas_version
+@pytest.mark.parametrize(
+    "column,expected",
+    [
+        ("base", 0),
+        ("floats", 0.2),
+        ("decimals", 0.1),
+        ("null_floats", 0.1),
+        ("strings", 0.1),
+        ("mixed_strings", 1),
+        ("infinity", np.inf),
+    ],
+)
+def test_calculate_max_diff(column, expected):
+    MAX_DIFF_DF = ps.DataFrame(
+        {
+            "base": [1, 1, 1, 1, 1],
+            "floats": [1.1, 1.1, 1.1, 1.2, 0.9],
+            "decimals": [
+                Decimal("1.1"),
+                Decimal("1.1"),
+                Decimal("1.1"),
+                Decimal("1.1"),
+                Decimal("1.1"),
+            ],
+            "null_floats": [np.nan, 1.1, 1, 1, 1],
+            "strings": ["1", "1", "1", "1.1", "1"],
+            "mixed_strings": ["1", "1", "1", "2", "some string"],
+            "infinity": [1, 1, 1, 1, np.inf],
+        }
     )
-    df2 = spark_session.createDataFrame(
-        [
-            (1, "foo", "test"),
-            (2, "baz", "test"),
-        ],
-        ["id", "例", "予測対象日"],
+    assert np.isclose(
+        calculate_max_diff(MAX_DIFF_DF["base"], MAX_DIFF_DF[column]), expected
     )
-    compare = SparkCompare(spark_session, df1, df2, join_columns=["例"])
+
+
+@pandas_version
+def test_dupes_with_nulls_strings():
+    df1 = ps.DataFrame(
+        {
+            "fld_1": [1, 2, 2, 3, 3, 4, 5, 5],
+            "fld_2": ["A", np.nan, np.nan, np.nan, np.nan, np.nan, np.nan, np.nan],
+            "fld_3": [1, 2, 2, 3, 3, 4, 5, 5],
+        }
+    )
+    df2 = ps.DataFrame(
+        {
+            "fld_1": [1, 2, 3, 4, 5],
+            "fld_2": ["A", np.nan, np.nan, np.nan, np.nan],
+            "fld_3": [1, 2, 3, 4, 5],
+        }
+    )
+    comp = SparkCompare(df1, df2, join_columns=["fld_1", "fld_2"])
+    assert comp.subset()
+
+
+@pandas_version
+def test_dupes_with_nulls_ints():
+    df1 = ps.DataFrame(
+        {
+            "fld_1": [1, 2, 2, 3, 3, 4, 5, 5],
+            "fld_2": [1, np.nan, np.nan, np.nan, np.nan, np.nan, np.nan, np.nan],
+            "fld_3": [1, 2, 2, 3, 3, 4, 5, 5],
+        }
+    )
+    df2 = ps.DataFrame(
+        {
+            "fld_1": [1, 2, 3, 4, 5],
+            "fld_2": [1, np.nan, np.nan, np.nan, np.nan],
+            "fld_3": [1, 2, 3, 4, 5],
+        }
+    )
+    comp = SparkCompare(df1, df2, join_columns=["fld_1", "fld_2"])
+    assert comp.subset()
+
+
+@pandas_version
+@pytest.mark.parametrize(
+    "dataframe,expected",
+    [
+        (ps.DataFrame({"a": [1, 2, 3], "b": [1, 2, 3]}), ps.Series([0, 0, 0])),
+        (
+            ps.DataFrame({"a": ["a", "a", "DATACOMPY_NULL"], "b": [1, 1, 2]}),
+            ps.Series([0, 1, 0]),
+        ),
+        (ps.DataFrame({"a": [-999, 2, 3], "b": [1, 2, 3]}), ps.Series([0, 0, 0])),
+        (
+            ps.DataFrame({"a": [1, np.nan, np.nan], "b": [1, 2, 2]}),
+            ps.Series([0, 0, 1]),
+        ),
+        (
+            ps.DataFrame({"a": ["1", np.nan, np.nan], "b": ["1", "2", "2"]}),
+            ps.Series([0, 0, 1]),
+        ),
+        (
+            ps.DataFrame(
+                {"a": [datetime(2018, 1, 1), np.nan, np.nan], "b": ["1", "2", "2"]}
+            ),
+            ps.Series([0, 0, 1]),
+        ),
+    ],
+)
+def test_generate_id_within_group(dataframe, expected):
+    assert (generate_id_within_group(dataframe, ["a", "b"]) == expected).all()
+
+
+@pandas_version
+def test_lower():
+    """This function tests the toggle to use lower case for column names or not"""
+    # should match
+    df1 = ps.DataFrame({"a": [1, 2, 3], "b": [0, 1, 2]})
+    df2 = ps.DataFrame({"a": [1, 2, 3], "B": [0, 1, 2]})
+    compare = SparkCompare(df1, df2, join_columns=["a"])
+    assert compare.matches()
+    # should not match
+    df1 = ps.DataFrame({"a": [1, 2, 3], "b": [0, 1, 2]})
+    df2 = ps.DataFrame({"a": [1, 2, 3], "B": [0, 1, 2]})
+    compare = SparkCompare(df1, df2, join_columns=["a"], cast_column_names_lower=False)
+    assert not compare.matches()
+
+    # test join column
+    # should match
+    df1 = ps.DataFrame({"a": [1, 2, 3], "b": [0, 1, 2]})
+    df2 = ps.DataFrame({"A": [1, 2, 3], "B": [0, 1, 2]})
+    compare = SparkCompare(df1, df2, join_columns=["a"])
+    assert compare.matches()
+    # should fail because "a" is not found in df2
+    df1 = ps.DataFrame({"a": [1, 2, 3], "b": [0, 1, 2]})
+    df2 = ps.DataFrame({"A": [1, 2, 3], "B": [0, 1, 2]})
+    expected_message = "df2 must have all columns from join_columns"
+    with raises(ValueError, match=expected_message):
+        compare = SparkCompare(
+            df1, df2, join_columns=["a"], cast_column_names_lower=False
+        )
+
+
+@pandas_version
+def test_integer_column_names():
+    """This function tests that integer column names would also work"""
+    df1 = ps.DataFrame({1: [1, 2, 3], 2: [0, 1, 2]})
+    df2 = ps.DataFrame({1: [1, 2, 3], 2: [0, 1, 2]})
+    compare = SparkCompare(df1, df2, join_columns=[1])
+    assert compare.matches()
+
+
+@pandas_version
+@mock.patch("datacompy.spark.render")
+def test_save_html(mock_render):
+    df1 = ps.DataFrame([{"a": 1, "b": 2}, {"a": 1, "b": 2}])
+    df2 = ps.DataFrame([{"a": 1, "b": 2}, {"a": 1, "b": 2}])
+    compare = SparkCompare(df1, df2, join_columns=["a"])
+
+    m = mock.mock_open()
+    with mock.patch("datacompy.spark.open", m, create=True):
+        # assert without HTML call
+        compare.report()
+        assert mock_render.call_count == 4
+        m.assert_not_called()
+
+    mock_render.reset_mock()
+    m = mock.mock_open()
+    with mock.patch("datacompy.spark.open", m, create=True):
+        # assert with HTML call
+        compare.report(html_file="test.html")
+        assert mock_render.call_count == 4
+        m.assert_called_with("test.html", "w")
+
+
+def test_pandas_version():
+    expected_message = "It seems like you are running Pandas 2+. Please note that Pandas 2+ will only be supported in Spark 4+. See: https://issues.apache.org/jira/browse/SPARK-44101. If you need to use Spark DataFrame with Pandas 2+ then consider using Fugue otherwise downgrade to Pandas 1.5.3"
+    df1 = ps.DataFrame([{"a": 1, "b": 2}, {"a": 1, "b": 2}])
+    df2 = ps.DataFrame([{"a": 1, "b": 2}, {"a": 1, "b": 2}])
+    with mock.patch("pandas.__version__", "2.0.0"):
+        with raises(Exception, match=re.escape(expected_message)):
+            SparkCompare(df1, df2, join_columns=["a"])
+
+    with mock.patch("pandas.__version__", "1.5.3"):
+        SparkCompare(df1, df2, join_columns=["a"])
+
+
+@pandas_version
+def test_unicode_columns():
+    df1 = ps.DataFrame(
+        [{"a": 1, "例": 2, "予測対象日": "test"}, {"a": 1, "例": 3, "予測対象日": "test"}]
+    )
+    df2 = ps.DataFrame(
+        [{"a": 1, "例": 2, "予測対象日": "test"}, {"a": 1, "例": 3, "予測対象日": "test"}]
+    )
+    compare = SparkCompare(df1, df2, join_columns=["例"])
+    assert compare.matches()
     # Just render the report to make sure it renders.
-    compare.report()
+    t = compare.report()
